@@ -22,7 +22,7 @@ keywords = [
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
 python-dotenv = {version = "^1.0.0"}
 click = {version = "^8.1.3"}
 PyYAML = {version = "^6.0.1"}
@@ -39,7 +39,7 @@ AMQPStorm = {version = "^2.10.4", optional = true}
 uuid = {version = "^1.30", optional = true}
 SQLAlchemy = {version = "^2.0.23", optional = true}
 gunicorn = {version = "^23.0.0", optional = true}
-uvicorn = {version = "^0.30.0", optional = true}
+uvicorn = {version = "^0.32.0", optional = true}
 tzlocal = {version = "^5.2", optional = true}
 bottle = {version = "^0.13.1", optional = true}
 pyfiglet = {version = "^1.0.2", optional = true}
@@ -50,9 +50,9 @@ jsonschema = {version = "^4.22.0", optional = true}
 snek-sploit = {version = "^0.8.3", optional = true}
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.4.3"
-pytest-cov = "^4.1.0"
-pytest-asyncio = "^0.23.3"
+pytest = "^8.3.3"
+pytest-cov = "^5.0.0"
+pytest-asyncio = "^0.24.0"
 pytest-django = "^4.7.0"
 pytest-mock = "^3.6.1"
 model-bakery = "^1.4.0"
