@@ -1,5 +1,4 @@
 import click
-from typing import Union
 
 from cryton.cli.utility import helpers
 from cryton.cli.utility.decorators import *
@@ -28,7 +27,7 @@ def step_list(
     offset: int,
     limit: int,
     localize: bool,
-    parameter_filters: tuple[tuple[str, Union[str, int]]],
+    parameter_filters: tuple[tuple[str, str | int]],
     parent: int,
 ) -> None:
     """
@@ -203,7 +202,7 @@ def step_execution_list(
     offset: int,
     limit: int,
     localize: bool,
-    parameter_filters: tuple[tuple[str, Union[str, int]]],
+    parameter_filters: tuple[tuple[str, str | int]],
     parent: int,
 ) -> None:
     """
@@ -314,21 +313,21 @@ def step_execution_report(ctx: helpers.Context, execution_id: int, file: str, le
     helpers.save_yaml(response, file, f"step-execution-{execution_id}.yml", less, less, localize, ctx.obj.debug)
 
 
-@step_execution.command("kill")
+@step_execution.command("stop")
 @click.pass_context
 @click.argument("execution_id", type=click.INT, required=True)
-def step_execution_kill(ctx: helpers.Context, execution_id: int) -> None:
+def step_execution_stop(ctx: helpers.Context, execution_id: int) -> None:
     """
-    Kill Step's execution with EXECUTION_ID saved in Cryton.
+    Stop Step's execution with EXECUTION_ID saved in Cryton.
 
-    EXECUTION_ID is ID of the Step's execution you want to kill.
+    EXECUTION_ID is ID of the Step's execution you want to stop.
 
     \f
     :param ctx: Click ctx object
     :param execution_id: ID of the desired Step's execution
     :return: None
     """
-    response = ctx.obj.api_post(StepExecution.KILL, execution_id)
+    response = ctx.obj.api_post(StepExecution.STOP, execution_id)
     helpers.print_message(response, ctx.obj.debug)
 
 
@@ -339,7 +338,7 @@ def step_execution_re_execute(ctx: helpers.Context, execution_id: int) -> None:
     """
     Re-execute Step's execution with EXECUTION_ID saved in Cryton.
 
-    EXECUTION_ID is ID of the Step's execution you want to kill.
+    EXECUTION_ID is ID of the Step's execution you want to re-execute.
 
     \f
     :param ctx: Click ctx object
