@@ -1,10 +1,8 @@
 import os
 from datetime import datetime
-from typing import Union, Type, Optional
+from typing import Type
 
-from django.core import exceptions as django_exc
 from django.db import transaction, connections
-from django.db.models.query import QuerySet
 from django.forms.models import model_to_dict
 
 from multiprocessing import Process
@@ -13,12 +11,13 @@ from cryton.hive.cryton_app.models import PlanModel, PlanExecutionModel, StageEx
 
 from cryton.hive.utility import constants, exceptions, logger, scheduler_client, states as st
 from cryton.hive.config.settings import SETTINGS
-from cryton.hive.models import worker
 from cryton.hive.models.stage import StageExecution
 from django.utils import timezone
+from cryton.hive.models.abstract import Instance, SchedulableExecution
+from cryton.hive.models.worker import Worker
 
 
-class Plan:
+class Plan(Instance):
     def __init__(self, model_id: int):
         """
         :param model_id: Model ID
@@ -38,7 +37,7 @@ class Plan:
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[PlanModel], PlanModel]:
+    def model(self) -> Type[PlanModel] | PlanModel:
         self.__model.refresh_from_db()
         return self.__model
 
@@ -105,30 +104,33 @@ class Plan:
         return plan_data
 
 
-class PlanExecution:
-    def __init__(self, **kwargs):
-        plan_execution_id = kwargs.get("plan_execution_id")
-        if plan_execution_id is not None:
-            try:
-                self.model = PlanExecutionModel.objects.get(id=plan_execution_id)
-            except django_exc.ObjectDoesNotExist:
-                raise exceptions.PlanExecutionDoesNotExist(plan_execution_id=plan_execution_id)
-        else:
-            self.model = PlanExecutionModel.objects.create(**kwargs)
-            self._prepare_executions()
-            self._create_evidence_directory()
+class PlanExecution(SchedulableExecution):
+    def __init__(self, model_id: int):
+        """
+        :param model_id: Model ID
+        """
+        self.__model = PlanExecutionModel.objects.get(id=model_id)
+
+    @staticmethod
+    def create_model(plan_id: int, worker_id: int, run_id: int) -> PlanExecutionModel | Type[PlanExecutionModel]:
+        return PlanExecutionModel.objects.create(plan_id=plan_id, worker_id=worker_id, run_id=run_id)
+
+    @classmethod
+    def prepare(cls, plan_id: int, worker_id: int, run_id: int) -> "PlanExecution":
+        model = cls.create_model(plan_id, worker_id, run_id)
+
+        for stage in model.plan.stages.all():
+            StageExecution.prepare(stage.id, model.id)
+
+        return PlanExecution(model.id)
 
     def delete(self):
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[PlanExecutionModel], PlanExecutionModel]:
+    def model(self) -> Type[PlanExecutionModel] | PlanExecutionModel:
         self.__model.refresh_from_db()
         return self.__model
-
-    @model.setter
-    def model(self, value: PlanExecutionModel):
-        self.__model = value
 
     @property
     def state(self) -> str:
@@ -138,7 +140,7 @@ class PlanExecution:
     def state(self, value: str):
         with transaction.atomic():
             PlanExecutionModel.objects.select_for_update().get(id=self.model.id)
-            if st.PlanStateMachine(self.model.id).validate_transition(self.state, value):
+            if st.PlanStateMachine.validate_transition(self.state, value):
                 logger.logger.debug(
                     "Plan execution changed state",
                     state_from=self.state,
@@ -150,41 +152,41 @@ class PlanExecution:
                 model.save()
 
     @property
-    def schedule_time(self) -> Optional[datetime]:
+    def schedule_time(self) -> datetime | None:
         return self.model.schedule_time
 
     @schedule_time.setter
-    def schedule_time(self, value: Optional[datetime]):
+    def schedule_time(self, value: datetime | None):
         model = self.model
         model.schedule_time = value
         model.save()
 
     @property
-    def start_time(self) -> Optional[datetime]:
+    def start_time(self) -> datetime | None:
         return self.model.start_time
 
     @start_time.setter
-    def start_time(self, value: Optional[datetime]):
+    def start_time(self, value: datetime | None):
         model = self.model
         model.start_time = value
         model.save()
 
     @property
-    def pause_time(self) -> Optional[datetime]:
+    def pause_time(self) -> datetime | None:
         return self.model.pause_time
 
     @pause_time.setter
-    def pause_time(self, value: Optional[datetime]):
+    def pause_time(self, value: datetime | None):
         model = self.model
         model.pause_time = value
         model.save()
 
     @property
-    def finish_time(self) -> Optional[datetime]:
+    def finish_time(self) -> datetime | None:
         return self.model.finish_time
 
     @finish_time.setter
-    def finish_time(self, value: Optional[datetime]):
+    def finish_time(self, value: datetime | None):
         model = self.model
         model.finish_time = value
         model.save()
@@ -213,28 +215,6 @@ class PlanExecution:
     def all_stages_finished(self) -> bool:
         return not self.model.stage_executions.all().exclude(state__in=st.STAGE_FINAL_STATES).exists()
 
-    def _create_evidence_directory(self) -> None:
-        """
-        Generate directory for storing execution evidence.
-        :return: None
-        """
-        worker_evidence_directory = os.path.join(
-            SETTINGS.evidence_directory, f"run_{self.model.run_id}", f"worker_{self.model.worker.name}"
-        )
-
-        os.makedirs(worker_evidence_directory, exist_ok=True)
-        self.evidence_directory = worker_evidence_directory
-
-    def _prepare_executions(self):
-        """
-        Prepare execution for each Stage.
-        :return: None
-        """
-        stage_execution_kwargs = {"plan_execution": self.model}
-        for stage_obj in self.model.plan.stages.all():
-            stage_execution_kwargs.update({"stage": stage_obj})
-            StageExecution(**stage_execution_kwargs)
-
     def schedule(self, schedule_time: datetime) -> None:
         """
         Schedule all plan's stages.
@@ -243,9 +223,8 @@ class PlanExecution:
         :raises
             :exception RuntimeError
         """
-        # Check state
         logger.logger.debug("Scheduling Plan execution", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_SCHEDULE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_SCHEDULE_STATES)
 
         self.aps_job_id = scheduler_client.schedule_function(
             "cryton.hive.models.plan:execution", [self.model.id], schedule_time
@@ -258,23 +237,26 @@ class PlanExecution:
         else:
             raise RuntimeError("Could not schedule Plan execution")
 
-    def execute(self) -> None:
+    def start(self) -> None:
         """
         Execute Plan. This method starts triggers.
         :return: None
         """
         logger.logger.debug("Executing Plan", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_EXECUTE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_EXECUTE_STATES)
+
+        # Create evidence directory
+        self.evidence_directory = os.path.join(
+            SETTINGS.evidence_directory, f"run_{self.model.run_id}", f"worker_{self.model.worker.name}"
+        )
+        os.makedirs(self.evidence_directory, exist_ok=True)
+        Worker(self.model.worker_id).prepare_rabbit_queues()
 
         self.start_time = timezone.now()
         self.state = st.RUNNING
 
-        # Prepare rabbit queues
-        worker_obj = worker.Worker(self.model.worker_id)
-        worker_obj.prepare_rabbit_queues()
-
-        # Start triggers
-        self.start_triggers()
+        for stage_execution_model in self.model.stage_executions.all():
+            StageExecution(stage_execution_model.id).start()
         logger.logger.info("Plan execution executed", plan_execution_id=self.model.id, status="success")
 
     def unschedule(self) -> None:
@@ -283,7 +265,7 @@ class PlanExecution:
         :return: None
         """
         logger.logger.debug("Unscheduling Plan execution", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_UNSCHEDULE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_UNSCHEDULE_STATES)
 
         scheduler_client.remove_job(self.aps_job_id)
         self.aps_job_id, self.schedule_time = "", None
@@ -298,7 +280,7 @@ class PlanExecution:
         :return: None
         """
         logger.logger.debug("Rescheduling Plan", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_RESCHEDULE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_RESCHEDULE_STATES)
 
         if new_time < timezone.now():
             raise exceptions.UserInputError("Time argument must be greater or equal than current time.", str(new_time))
@@ -313,37 +295,37 @@ class PlanExecution:
         :return: None
         """
         logger.logger.debug("Pausing Plan", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_PAUSE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_PAUSE_STATES)
 
         self.state = st.PAUSING
         logger.logger.info("Plan execution pausing", plan_execution_id=self.model.id, status="success")
 
         for stage_ex in self.model.stage_executions.all():  # Pause StageExecutions.
-            StageExecution(stage_execution_id=stage_ex.id).trigger.pause()
+            StageExecution(stage_ex.id).pause()
 
         if not self.model.stage_executions.exclude(state__in=st.PLAN_STAGE_PAUSE_STATES).exists():
             self.state = st.PAUSED
             self.pause_time = timezone.now()
             logger.logger.info("Plan execution paused", plan_execution_id=self.model.id, status="success")
 
-    def unpause(self) -> None:
+    def resume(self) -> None:
         logger.logger.debug("Unpausing Plan execution", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_UNPAUSE_STATES)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_RESUME_STATES)
 
         self.pause_time = None
         self.state = st.RUNNING
 
         for stage_execution_model in self.model.stage_executions.all():
-            stage_ex = StageExecution(stage_execution_id=stage_execution_model.id)
-            if stage_ex.model.stage.type == constants.DELTA and stage_ex.state in st.STAGE_SCHEDULE_STATES:
-                stage_ex.trigger.start()
+            stage_ex = StageExecution(stage_execution_model.id)
+            if stage_ex.model.stage.type == constants.DELTA and stage_ex.state in [st.AWAITING]:
+                stage_ex.start()
             else:
                 try:
-                    stage_ex.trigger.unpause()
-                except exceptions.StageInvalidStateError:
+                    stage_ex.resume()
+                except exceptions.InvalidStateError:
                     pass
 
-        logger.logger.info("Plan execution unpaused", plan_execution_id=self.model.id)
+        logger.logger.info("Plan execution resumed", plan_execution_id=self.model.id)
 
     def validate_modules(self):
         """
@@ -352,41 +334,9 @@ class PlanExecution:
         logger.logger.debug("Validating Plan execution modules", plan_execution_id=self.model.id)
 
         for stage_execution_id in self.model.stage_executions.values_list("id", flat=True):
-            stage_execution = StageExecution(stage_execution_id=stage_execution_id)
+            stage_execution = StageExecution(stage_execution_id)
             stage_execution.validate_modules()
         logger.logger.info("Plan execution modules validated", plan_execution_id=self.model.id)
-
-        return None
-
-    def start_triggers(self) -> None:
-        """
-        Start triggers for all execution stages.
-        :return: None
-        """
-        logger.logger.debug("Starting triggers", plan_execution_id=self.model.id)
-        for stage_execution_model in self.model.stage_executions.all():
-            stage_execution = StageExecution(stage_execution_id=stage_execution_model.id)
-            stage_execution.trigger.start()
-            logger.logger.debug(
-                "Trigger started", plan_execution_id=self.model.id, trigger=str(stage_execution.trigger)
-            )
-        logger.logger.info("Triggers started", plan_execution_id=self.model.id, status="success")
-
-    @staticmethod
-    def filter(**kwargs) -> QuerySet:
-        """
-        List PlanExecutionModel objects fulfilling fields specified in kwargs.
-        If no such fields are specified all objects are returned.
-        :param kwargs: dict of field-value pairs to filter by
-        :return: Queryset of PlanExecutionModel objects
-        :raises WrongParameterError: invalid field is specified
-        """
-        if kwargs:
-            try:
-                return PlanExecutionModel.objects.filter(**kwargs)
-            except django_exc.FieldError as ex:
-                raise exceptions.WrongParameterError(message=ex)
-        return PlanExecutionModel.objects.all()
 
     def report(self) -> dict:
         """
@@ -394,51 +344,48 @@ class PlanExecution:
         :return: report from Plan execution
         """
         logger.logger.debug("Generating report", plan_execution_id=self.model.id)
-
-        report_dict = dict()
-        report_dict.update(
-            {
-                "id": self.model.id,
-                "plan_name": self.model.plan.name,
-                "metadata": self.model.plan.metadata,
-                "state": self.state,
-                "schedule_time": self.schedule_time,
-                "start_time": self.start_time,
-                "finish_time": self.finish_time,
-                "pause_time": self.pause_time,
-                "worker_id": self.model.worker_id,
-                "worker_name": self.model.worker.name,
-                "evidence_directory": self.evidence_directory,
-            }
+        report_dict = dict(
+            id=self.model.id,
+            plan_name=self.model.plan.name,
+            metadata=self.model.plan.metadata,
+            state=self.state,
+            schedule_time=self.schedule_time,
+            start_time=self.start_time,
+            finish_time=self.finish_time,
+            pause_time=self.pause_time,
+            worker_id=self.model.worker_id,
+            worker_name=self.model.worker.name,
+            evidence_directory=self.evidence_directory,
+            stage_executions=[],
         )
-        report_dict.update({"stage_executions": []})
+
         for stage_execution_obj in StageExecutionModel.objects.filter(plan_execution_id=self.model.id).order_by("id"):
-            stage_ex_report = StageExecution(stage_execution_id=stage_execution_obj.id).report()
+            stage_ex_report = StageExecution(stage_execution_obj.id).report()
             report_dict["stage_executions"].append(stage_ex_report)
 
         return report_dict
 
-    def kill(self) -> None:
+    def stop(self) -> None:
         """
-        Kill current PlanExecution and its StageExecutions
+        Stop current PlanExecution and its StageExecutions
         :return: None
         """
-        logger.logger.debug("Killing Plan", plan_execution_id=self.model.id)
-        st.PlanStateMachine(self.model.id).validate_state(self.state, st.PLAN_KILL_STATES)
-        self.state = st.TERMINATING
+        logger.logger.debug("Stoping Plan", plan_execution_id=self.model.id)
+        st.PlanStateMachine.validate_state(self.state, st.PLAN_STOP_STATES)
+        self.state = st.STOPPING
 
         processes = list()
 
         # Unschedule Stage executions that can be unscheduled
-        for stage_ex_model in self.model.stage_executions.filter(state__in=st.STAGE_UNSCHEDULE_STATES):
-            stage_ex = StageExecution(stage_execution_id=stage_ex_model.id)
+        for stage_ex_model in self.model.stage_executions.filter(state=st.AWAITING):
+            stage_ex = StageExecution(stage_ex_model.id)
             process = Process(target=stage_ex.trigger.stop)
             processes.append(process)
 
-        # Kill Stage executions that cannot be unscheduled
-        for stage_ex_model in self.model.stage_executions.filter(state__in=st.STAGE_KILL_STATES):
-            stage_ex = StageExecution(stage_execution_id=stage_ex_model.id)
-            process = Process(target=stage_ex.kill)
+        # Stop Stage executions that cannot be unscheduled
+        for stage_ex_model in self.model.stage_executions.filter(state__in=st.STAGE_STOP_STATES):
+            stage_ex = StageExecution(stage_ex_model.id)
+            process = Process(target=stage_ex.stop)
             processes.append(process)
 
         connections.close_all()  # close connections to force each process to create its own
@@ -449,19 +396,15 @@ class PlanExecution:
             process.join()
 
         self.finish_time = timezone.now()
-        self.state = st.TERMINATED
-        logger.logger.info(
-            "Plan execution killed", plan_execution_id=self.model.id, plan_id=self.model.plan.id, status="success"
-        )
-
-        return None
+        self.state = st.STOPPED
+        logger.logger.info("Plan execution stoped", plan_execution_id=self.model.id, plan_id=self.model.plan.id)
 
     def finish(self) -> None:
         """
         Finish execution - update necessary variables.
         :return: None
         """
-        st.PlanStateMachine(self.model.id).validate_state(self.state, [st.PAUSING, st.RUNNING])
+        st.PlanStateMachine.validate_state(self.state, [st.PAUSING, st.RUNNING])
         logger.logger.info("Plan execution finished", plan_execution_id=self.model.id)
 
         self.state = st.FINISHED
@@ -474,5 +417,4 @@ def execution(plan_execution_id: int) -> None:
     :param plan_execution_id: desired PlanExecutionModel's ID
     :return: None
     """
-    PlanExecution(plan_execution_id=plan_execution_id).execute()
-    return None
+    PlanExecution(plan_execution_id).start()
