@@ -1,123 +1,94 @@
-from typing import List, Optional, Union, Type
-from datetime import datetime, timedelta
+from typing import Type
+from datetime import datetime
 from threading import Thread
 
-from django.db.models.query import QuerySet
-from django.core import exceptions as django_exc
 from django.db import transaction
 from django.utils import timezone
 
-from cryton.hive.cryton_app.models import RunModel, PlanExecutionModel, WorkerModel
-from cryton.hive.utility import exceptions, logger, scheduler_client, states as st
+from cryton.hive.cryton_app.models import RunModel
+from cryton.hive.utility import logger, scheduler_client, states as st
 from cryton.hive.models.plan import PlanExecution
 from cryton.hive.models.worker import Worker
-
-from dataclasses import dataclass, asdict
-
-
-@dataclass
-class RunReport:
-    id: int
-    plan_id: int
-    plan_name: str
-    state: str
-    schedule_time: datetime
-    start_time: datetime
-    pause_time: datetime
-    finish_time: datetime
-    plan_executions: list
+from cryton.hive.models.abstract import SchedulableExecution
 
 
-class Run:
-    def __init__(self, **kwargs):  # TODO: update
+class Run(SchedulableExecution):
+    def __init__(self, model_id: int):
         """
-        :param kwargs:
-            run_id: int = None,
-            plan_id: int
-            workers: list
+        :param model_id: Model ID
         """
-        if kwargs.get("run_model_id"):
-            run_id = kwargs.get("run_model_id")
+        self.__model = RunModel.objects.get(id=model_id)
 
-            try:
-                self.model = RunModel.objects.get(id=run_id)
-            except django_exc.ObjectDoesNotExist:
-                raise exceptions.RunObjectDoesNotExist(run_id=run_id)
-            self.workers = [pex.worker for pex in self.model.plan_executions.all()]
-        else:
-            workers: list = kwargs.pop("workers", [])
-            if not workers:
-                raise exceptions.WrongParameterError(message="Parameter cannot be empty.", param_name="workers")
+    @staticmethod
+    def create_model(plan_id: int) -> RunModel | Type[RunModel]:
+        return RunModel.objects.create(plan_id=plan_id)
 
-            self.workers = workers
-            self.model = RunModel.objects.create(**kwargs)
-            self._prepare_executions()
+    @classmethod
+    def prepare(cls, plan_id: int, worker_ids: list[int]) -> "Run":
+        model = cls.create_model(plan_id)
+        for worker_id in worker_ids:
+            PlanExecution.prepare(plan_id, worker_id, model.id)
+
+        return Run(model.id)
 
     def delete(self):
         """
         Delete RunModel
         :return:
         """
-        logger.logger.debug("Deleting run", run_id=self.model.id)
+        logger.logger.debug("deleting run", run_id=self.model.id)
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[RunModel], RunModel]:
-        """
-        Get or set the RunModel.
-        """
+    def model(self) -> Type[RunModel] | RunModel:
         self.__model.refresh_from_db()
         return self.__model
 
-    @model.setter
-    def model(self, value):
-        self.__model = value
-
     @property
-    def schedule_time(self) -> Optional[datetime]:
+    def schedule_time(self) -> datetime | None:
         return self.model.schedule_time
 
     @schedule_time.setter
-    def schedule_time(self, value: Optional[datetime]):
+    def schedule_time(self, value: datetime | None):
         model = self.model
         model.schedule_time = value
         model.save()
 
     @property
-    def start_time(self) -> Optional[datetime]:
+    def start_time(self) -> datetime | None:
         """
         Get or set start time of RunModel.
         """
         return self.model.start_time
 
     @start_time.setter
-    def start_time(self, value: Optional[datetime]):
+    def start_time(self, value: datetime | None):
         model = self.model
         model.start_time = value
         model.save()
 
     @property
-    def pause_time(self) -> Optional[datetime]:
+    def pause_time(self) -> datetime | None:
         """
         Get or set pause time of RunModel.
         """
         return self.model.pause_time
 
     @pause_time.setter
-    def pause_time(self, value: Optional[datetime]):
+    def pause_time(self, value: datetime | None):
         model = self.model
         model.pause_time = value
         model.save()
 
     @property
-    def finish_time(self) -> Optional[datetime]:
+    def finish_time(self) -> datetime | None:
         """
         Get or set pause time of RunModel.
         """
         return self.model.finish_time
 
     @finish_time.setter
-    def finish_time(self, value: Optional[datetime]):
+    def finish_time(self, value: datetime | None):
         model = self.model
         model.finish_time = value
         model.save()
@@ -133,7 +104,7 @@ class Run:
     def state(self, value: str):
         with transaction.atomic():
             RunModel.objects.select_for_update().get(id=self.model.id)
-            if st.RunStateMachine(self.model.id).validate_transition(self.state, value):
+            if st.RunStateMachine.validate_transition(self.state, value):
                 logger.logger.debug("run changed state", state_from=self.state, state_to=value, run_id=self.model.id)
                 model = self.model
                 model.state = value
@@ -150,45 +121,11 @@ class Run:
         model.save()
 
     @property
-    def workers(self) -> List[WorkerModel]:
-        return self.__workers
-
-    @workers.setter
-    def workers(self, value: List[WorkerModel]):
-        self.__workers = value
-
-    @property
     def all_plans_finished(self) -> bool:
         return not self.model.plan_executions.all().exclude(state__in=st.PLAN_FINAL_STATES).exists()
 
-    @staticmethod
-    def filter(**kwargs) -> QuerySet:
-        """
-        Get list of RunModel according to no or specified conditions
-        :param kwargs: dict of parameters to filter by
-        :return:
-        """
-        logger.logger.debug("Filtering runs")
-        if kwargs:
-            try:
-                return RunModel.objects.filter(**kwargs)
-            except django_exc.FieldError as ex:
-                raise exceptions.WrongParameterError(message=ex)
-        else:
-            return RunModel.objects.all()
-
-    def _prepare_executions(self):
-        """
-        Prepare execution for each worker.
-        :return: None
-        """
-        plan_execution_kwargs = {"run": self.model, "plan_id": self.model.plan.id}
-        for worker_obj in self.workers:
-            plan_execution_kwargs.update({"worker": worker_obj})
-            PlanExecution(**plan_execution_kwargs)
-
     def report(self) -> dict:
-        report_obj = RunReport(
+        report_obj = dict(
             id=self.model.id,
             plan_id=self.model.plan.id,
             plan_name=self.model.plan.name,
@@ -200,11 +137,11 @@ class Run:
             plan_executions=[],
         )
 
-        for plan_execution_obj in PlanExecutionModel.objects.filter(run_id=self.model.id).order_by("id"):
-            plan_ex_report = PlanExecution(plan_execution_id=plan_execution_obj.id).report()
-            report_obj.plan_executions.append(plan_ex_report)
+        for plan_execution_model in self.model.plan_executions.order_by("id"):
+            plan_execution_report = PlanExecution(plan_execution_model.id).report()
+            report_obj["plan_executions"].append(plan_execution_report)
 
-        return asdict(report_obj)
+        return report_obj
 
     def schedule(self, schedule_time: datetime) -> None:
         """
@@ -214,11 +151,9 @@ class Run:
         :raises
             :exception RuntimeError
         """
-        logger.logger.debug("Scheduling Run", run_id=self.model.id)
-        # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_SCHEDULE_STATES)
+        logger.logger.debug("scheduling run", run_id=self.model.id)
+        st.RunStateMachine.validate_state(self.state, st.RUN_SCHEDULE_STATES)
 
-        # Schedule execution
         self.aps_job_id = scheduler_client.schedule_function(
             "cryton.hive.models.run:execution", [self.model.id], schedule_time
         )
@@ -236,7 +171,7 @@ class Run:
         """
         logger.logger.debug("Unscheduling Run", run_id=self.model.id)
         # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_UNSCHEDULE_STATES)
+        st.RunStateMachine.validate_state(self.state, st.RUN_UNSCHEDULE_STATES)
 
         scheduler_client.remove_job(self.aps_job_id)
         self.aps_job_id, self.schedule_time = "", None
@@ -251,7 +186,7 @@ class Run:
         """
         logger.logger.debug("Rescheduling Run", run_id=self.model.id)
         # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_RESCHEDULE_STATES)
+        st.RunStateMachine.validate_state(self.state, st.RUN_RESCHEDULE_STATES)
 
         self.unschedule()
         self.schedule(schedule_time)
@@ -265,51 +200,34 @@ class Run:
         """
         logger.logger.debug("Pausing Run", run_id=self.model.id)
         # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_PAUSE_STATES)
+        st.RunStateMachine.validate_state(self.state, st.RUN_PAUSE_STATES)
 
         self.state = st.PAUSING
 
         for plan_ex in self.model.plan_executions.all():
-            PlanExecution(plan_execution_id=plan_ex.id).pause()
+            PlanExecution(plan_ex.id).pause()
 
         if not self.model.plan_executions.exclude(state__in=st.RUN_PLAN_PAUSE_STATES).exists():
             self.state = st.PAUSED
             self.pause_time = timezone.now()
             logger.logger.info("Run paused", run_id=self.model.id, status="success")
 
-    def unpause(self) -> None:
+    def resume(self) -> None:
         """
-        Unpauses Run on specified WorkerModels
+        Resumes Run on specified WorkerModels
         :return:
         """
         logger.logger.debug("Unpausing Run", run_id=self.model.id)
         # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_UNPAUSE_STATES)
+        st.RunStateMachine.validate_state(self.state, st.RUN_RESUME_STATES)
 
         self.pause_time = None
         self.state = st.RUNNING
 
         for plan_execution_model in self.model.plan_executions.all():
-            PlanExecution(plan_execution_id=plan_execution_model.id).unpause()
+            PlanExecution(plan_execution_model.id).resume()
 
-        logger.logger.info("Run unpaused", run_id=self.model.id, status="success")
-
-    def postpone(self, delta: timedelta) -> None:
-        """
-        Postpones Run on specified WorkerModels
-        :param delta: Time delta
-        :return:
-        """
-        logger.logger.debug("Postponing Run", run_id=self.model.id)
-        # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_POSTPONE_STATES)
-
-        schedule_time = self.schedule_time + delta
-
-        self.unschedule()
-        self.schedule(schedule_time)
-
-        logger.logger.info("Run postponed", run_id=self.model.id, status="success")
+        logger.logger.info("Run resumed", run_id=self.model.id, status="success")
 
     def healthcheck_workers(self) -> None:
         """
@@ -317,64 +235,57 @@ class Run:
         :raise ConnectionError: If a Worker is down.
         :return: None
         """
-        for worker_model in self.workers:
-            worker = Worker(worker_model.id)
+        for plan_ex_model in self.model.plan_executions.all():
+            worker = Worker(plan_ex_model.worker.id)
             if not worker.healthcheck():
                 raise ConnectionError(f"Unable to get response from worker {worker.name} ({worker.model.id}).")
 
-    def execute(self) -> None:
+    def start(self) -> None:
         """
         Executes Run
         :return:
         """
         logger.logger.debug("Executing Run", run_id=self.model.id)
-        # Check state
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_EXECUTE_STATES)
+        st.RunStateMachine.validate_state(self.state, st.RUN_EXECUTE_STATES)
 
         self.start_time = timezone.now()
         self.state = st.RUNNING
 
-        # Execute Plan on every Slave
-        for worker_obj in self.workers:
-            plan_execution_model = self.model.plan_executions.get(worker_id=worker_obj.id)
-            PlanExecution(plan_execution_id=plan_execution_model.id).execute()
+        for plan_ex_model in self.model.plan_executions.all():
+            PlanExecution(plan_ex_model.id).start()
 
-        logger.logger.info("Run executed", run_id=self.model.id, status="success")
+        logger.logger.info("Run executed", run_id=self.model.id)
 
-    def kill(self) -> None:
+    def stop(self) -> None:
         """
-        Kill current Run and its PlanExecutions
+        Stop current Run and its PlanExecutions
         :return: None
         """
-        logger.logger.debug("Killing Run", run_id=self.model.id)
-        st.RunStateMachine(self.model.id).validate_state(self.state, st.RUN_KILL_STATES)
-        self.state = st.TERMINATING
+        logger.logger.debug("stopping run", run_id=self.model.id)
+        st.RunStateMachine.validate_state(self.state, st.RUN_STOP_STATES)
+        self.state = st.STOPPING
 
         threads = list()
-
         # Unschedule Plan executions that can be unscheduled
         for plan_ex_obj in self.model.plan_executions.filter(state__in=st.PLAN_UNSCHEDULE_STATES):
-            plan_ex = PlanExecution(plan_execution_id=plan_ex_obj.id)
+            plan_ex = PlanExecution(plan_ex_obj.id)
             thread = Thread(target=plan_ex.unschedule)
-            threads.append(thread)
-
-        # Kill Plan executions that cannot be unscheduled
-        for plan_ex_obj in self.model.plan_executions.filter(state__in=st.PLAN_KILL_STATES):
-            plan_ex = PlanExecution(plan_execution_id=plan_ex_obj.id)
-            thread = Thread(target=plan_ex.kill)
-            threads.append(thread)
-
-        for thread in threads:
             thread.start()
+            threads.append(thread)
+
+        # Stop Plan executions that cannot be unscheduled
+        for plan_ex_obj in self.model.plan_executions.filter(state__in=st.PLAN_STOP_STATES):
+            plan_ex = PlanExecution(plan_ex_obj.id)
+            thread = Thread(target=plan_ex.stop)
+            thread.start()
+            threads.append(thread)
 
         for thread in threads:
             thread.join()
 
         self.finish_time = timezone.now()
-        self.state = st.TERMINATED
-        logger.logger.info("Run killed", run_id=self.model.id, status="success")
-
-        return None
+        self.state = st.STOPPED
+        logger.logger.info("run stopped", run_id=self.model.id)
 
     def validate_modules(self):
         """
@@ -383,18 +294,16 @@ class Run:
         logger.logger.debug("Run modules validation started", run_id=self.model.id)
 
         for plan_execution_id in self.model.plan_executions.values_list("id", flat=True):
-            plan_execution = PlanExecution(plan_execution_id=plan_execution_id)
+            plan_execution = PlanExecution(plan_execution_id)
             plan_execution.validate_modules()
         logger.logger.debug("Run modules validated", run_id=self.model.id)
-
-        return None
 
     def finish(self) -> None:
         """
         Finish execution - update necessary variables.
         :return: None
         """
-        st.RunStateMachine(self.model.id).validate_state(self.state, [st.PAUSING, st.RUNNING])
+        st.RunStateMachine.validate_state(self.state, [st.PAUSING, st.RUNNING])
         logger.logger.info("Run finished", run_id=self.model.id)
 
         self.state = st.FINISHED
@@ -407,5 +316,4 @@ def execution(run_model_id: int) -> None:
     :param run_model_id: desired RunModel's ID
     :return: None
     """
-    Run(run_model_id=run_model_id).execute()
-    return None
+    Run(run_model_id).start()
