@@ -1,35 +1,27 @@
-from typing import Union, Type, Optional, List
+from typing import Type
 from datetime import datetime
 from django.utils import timezone
 from threading import Thread
 
-from django.db.models.query import QuerySet
-from django.core import exceptions as django_exc
 from django.db import transaction, connections
 
-from cryton.hive.cryton_app.models import StageModel, StageExecutionModel, StepExecutionModel, StageDependencyModel
-from cryton.hive.utility import exceptions, logger, states as st, util
-from cryton.hive.triggers import TriggerType, TriggerDelta, TriggerHTTP, TriggerMSF, TriggerDateTime
-from cryton.hive.models.step import StepExecution, Step
+from cryton.hive.cryton_app.models import StageModel, StageExecutionModel
+from cryton.hive.utility import logger, states as st, util
+from cryton.hive.triggers import (
+    TriggerType,
+    TriggerDelta,
+    TriggerHTTP,
+    TriggerMetasploit,
+    TriggerTime,
+    TriggerImmediate,
+)
+from cryton.hive.models.step import StepExecution
+from cryton.hive.models.worker import Worker
 from cryton.hive.config.settings import SETTINGS
-
-from dataclasses import dataclass, asdict
-
-
-@dataclass
-class StageReport:
-    id: int
-    name: str
-    metadata: dict
-    state: str
-    start_time: datetime
-    pause_time: datetime
-    finish_time: datetime
-    schedule_time: datetime
-    step_executions: list
+from cryton.hive.models.abstract import Instance, Execution
 
 
-class Stage:
+class Stage(Instance):
     def __init__(self, model_id: int):
         """
         :param model_id: Model ID
@@ -46,7 +38,7 @@ class Stage:
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[StageModel], StageModel]:
+    def model(self) -> Type[StageModel] | StageModel:
         self.__model.refresh_from_db()
         return self.__model
 
@@ -54,104 +46,42 @@ class Stage:
     def name(self) -> str:
         return self.model.name
 
-    @name.setter
-    def name(self, value):
-        model = self.model
-        model.name = value
-        model.save()
-
     @property
     def trigger_type(self) -> str:
         return self.model.type
-
-    @trigger_type.setter
-    def trigger_type(self, value: str):
-        model = self.model
-        model.type = value
-        model.save()
 
     @property
     def meta(self) -> dict:
         return self.model.metadata
 
-    @meta.setter
-    def meta(self, value: dict):
-        model = self.model
-        model.metadata = value
-        model.save()
 
-    @property
-    def final_steps(self) -> QuerySet:
-        steps_list = Step.filter(stage_id=self.model.id, is_final=True)
-        return steps_list
-
-    @property
-    def execution_list(self) -> QuerySet:
+class StageExecution(Execution):
+    def __init__(self, model_id: int):
         """
-        Returns StageExecutionModel QuerySet. If the latest is needed, use '.latest()' on result.
-        :return: QuerySet of StageExecutionModel
+        :param model_id: Model ID
         """
-        return StageExecutionModel.objects.filter(stage_id=self.model.id)
+        self.__model = StageExecutionModel.objects.get(id=model_id)
 
     @staticmethod
-    def filter(**kwargs) -> QuerySet:
-        """
-        :param kwargs: dict of parameters to filter by
-        :return: QuerySet of StageModel
-        """
-        if kwargs:
-            try:
-                return StageModel.objects.filter(**kwargs)
-            except django_exc.FieldError as e:
-                raise exceptions.WrongParameterError(e)
-        else:
-            return StageModel.objects.all()
+    def create_model(stage_id: int, plan_execution_id: int) -> StageExecutionModel | Type[StageExecutionModel]:
+        return StageExecutionModel.objects.create(stage_id=stage_id, plan_execution_id=plan_execution_id)
 
-    def add_dependency(self, dependency_id: int) -> int:
-        """
-        Create dependency object
-        :param dependency_id: Stage ID
-        :return: ID of the dependency object
-        """
-        logger.logger.debug("Creating stage dependency", stage_id=self.model.id, dependency_id=dependency_id)
-        dependency_obj = StageDependencyModel(stage_id=self.model.id, dependency_id=dependency_id)
-        dependency_obj.save()
-        logger.logger.debug("Stage dependency created", stage_id=self.model.id, dependency_id=dependency_id)
+    @classmethod
+    def prepare(cls, stage_id: int, plan_execution_id: int) -> "StageExecution":
+        model = cls.create_model(stage_id, plan_execution_id)
 
-        return dependency_obj.id
+        for step in model.stage.steps.all():
+            StepExecution.prepare(step.id, model.id)
 
-
-class StageExecution:
-    def __init__(self, **kwargs):
-        """
-        :param kwargs:
-        (optional) stage_execution_id: int - for retrieving existing execution
-        stage_id: int - for creating new execution
-        """
-        stage_execution_id = kwargs.get("stage_execution_id")
-        if stage_execution_id:
-            try:
-                self.model = StageExecutionModel.objects.get(id=stage_execution_id)
-            except django_exc.ObjectDoesNotExist:
-                raise exceptions.StageExecutionObjectDoesNotExist(
-                    "StageExecutionModel with id {} does not exist.".format(stage_execution_id), stage_execution_id
-                )
-
-        else:
-            self.model = StageExecutionModel.objects.create(**kwargs)
-            self._prepare_executions()
+        return StageExecution(model.id)
 
     def delete(self):
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[StageExecutionModel], StageExecutionModel]:
+    def model(self) -> Type[StageExecutionModel] | StageExecutionModel:
         self.__model.refresh_from_db()
         return self.__model
-
-    @model.setter
-    def model(self, value: StageExecutionModel):
-        self.__model = value
 
     @property
     def state(self) -> str:
@@ -161,7 +91,7 @@ class StageExecution:
     def state(self, value: str):
         with transaction.atomic():
             StageExecutionModel.objects.select_for_update().get(id=self.model.id)
-            if st.StageStateMachine(self.model.id).validate_transition(self.state, value):
+            if st.StageStateMachine.validate_transition(self.state, value):
                 logger.logger.debug(
                     "Stage execution changed state",
                     state_from=self.state,
@@ -183,41 +113,41 @@ class StageExecution:
         model.save()
 
     @property
-    def start_time(self) -> Optional[datetime]:
+    def start_time(self) -> datetime | None:
         return self.model.start_time
 
     @start_time.setter
-    def start_time(self, value: Optional[datetime]):
+    def start_time(self, value: datetime | None):
         model = self.model
         model.start_time = value
         model.save()
 
     @property
-    def schedule_time(self) -> Optional[datetime]:
+    def schedule_time(self) -> datetime | None:
         return self.model.schedule_time
 
     @schedule_time.setter
-    def schedule_time(self, value: Optional[datetime]):
+    def schedule_time(self, value: datetime | None):
         model = self.model
         model.schedule_time = value
         model.save()
 
     @property
-    def pause_time(self) -> Optional[datetime]:
+    def pause_time(self) -> datetime | None:
         return self.model.pause_time
 
     @pause_time.setter
-    def pause_time(self, value: Optional[datetime]):
+    def pause_time(self, value: datetime | None):
         model = self.model
         model.pause_time = value
         model.save()
 
     @property
-    def finish_time(self) -> Optional[datetime]:
+    def finish_time(self) -> datetime | None:
         return self.model.finish_time
 
     @finish_time.setter
-    def finish_time(self, value: Optional[datetime]):
+    def finish_time(self, value: datetime | None):
         model = self.model
         model.finish_time = value
         model.save()
@@ -243,19 +173,18 @@ class StageExecution:
         model.save()
 
     @property
-    def serialized_output(self) -> Union[list, dict]:
+    def serialized_output(self) -> list | dict:
         return self.model.serialized_output
 
     @serialized_output.setter
-    def serialized_output(self, value: Union[list, dict]):
+    def serialized_output(self, value: list | dict):
         model = self.model
         model.serialized_output = value
         model.save()
 
     @property
-    def trigger(self) -> Union[TriggerDelta, TriggerHTTP, TriggerMSF, TriggerDateTime]:
-        trigger_type = self.model.stage.type
-        return TriggerType[trigger_type].value(stage_execution=self)
+    def trigger(self) -> TriggerDelta | TriggerHTTP | TriggerMetasploit | TriggerTime | TriggerImmediate:
+        return TriggerType[self.model.stage.type].value()
 
     @property
     def all_steps_finished(self) -> bool:
@@ -263,42 +192,17 @@ class StageExecution:
 
     @property
     def all_dependencies_finished(self) -> bool:
-        dependency_ids = self.model.stage.dependencies.all().values_list("dependency_id", flat=True)
-        cond = (
-            self.filter(stage_id__in=dependency_ids, plan_execution_id=self.model.plan_execution_id)
-            .exclude(state=st.FINISHED)
-            .exists()
-        )
+        dependencies = self.model.stage.dependencies.values("dependency")
+        execution_dependencies = self.model.plan_execution.stage_executions.filter(stage__in=dependencies)
+        unfinished_dependencies = execution_dependencies.exclude(state=st.FINISHED)
+        return not unfinished_dependencies.exists()
 
-        return not cond
-
-    @staticmethod
-    def filter(**kwargs) -> QuerySet:
-        """
-        Get list of StageExecutionModel according to no or specified conditions
-        :param kwargs: dict of parameters to filter by
-        :return: Desired QuerySet
-        """
-        if kwargs:
-            try:
-                return StageExecutionModel.objects.filter(**kwargs)
-            except django_exc.FieldError as ex:
-                raise exceptions.WrongParameterError(message=ex)
-        else:
-            return StageExecutionModel.objects.all()
-
-    def _prepare_executions(self):
-        """
-        Prepare execution for each Step.
-        :return: None
-        """
-        step_execution_kwargs = {"stage_execution": self.model}
-        for step_obj in self.model.stage.steps.all():
-            step_execution_kwargs.update({"step": step_obj})
-            StepExecution(**step_execution_kwargs)
+    @property
+    def control_queue(self):
+        return Worker(self.model.plan_execution.worker.id).control_queue
 
     @staticmethod
-    def _run_step_executions(step_executions: List[StepExecution]):
+    def _run_step_executions(step_executions: list[StepExecution]):
         """
         Evenly distribute Step executions and run each batch in a Process.
         :param step_executions: Step executions to be distributed into Processes
@@ -322,13 +226,62 @@ class StageExecution:
         for process in processes:
             process.start()
 
+    def start(self):
+        self.state = st.STARTING
+
+        try:
+            trigger_id, schedule_time = self.trigger.start(
+                arguments=self.model.stage.arguments, stage_execution_id=self.model.id, queue=self.control_queue
+            )
+        except Exception as ex:
+            # TODO: When changing the Stage's state to final state, it must propagate to PlanEx
+            #  Each execution component should have its `process_finished` method, maybe create event `process_finished`
+            logger.logger.error(f"Unable to start trigger", stage_execution_id=self.model.id, error=str(ex))
+            self.state = st.ERROR
+            return
+
+        if schedule_time:
+            self.schedule_time = schedule_time
+        if trigger_id:
+            self.trigger_id = trigger_id
+        self.state = st.AWAITING
+
+        if isinstance(self.trigger, TriggerImmediate):
+            Thread(target=self.execute).start()
+
+    def stop(self):
+        logger.logger.debug("stopping stage", stage_execution_id=self.model.id)
+        st.StageStateMachine.validate_state(self.state, st.STAGE_STOP_STATES)
+
+        state_before = self.state
+        self.state = st.STOPPING
+
+        if state_before == st.AWAITING:
+            self.trigger.stop(trigger_id=self.trigger_id, queue=self.control_queue)
+        elif state_before == st.WAITING:
+            pass
+        else:
+            threads = list()
+            for step_ex_model in self.model.step_executions.filter(state__in=st.STEP_STOP_STATES):
+                step_ex = StepExecution(step_ex_model.id)
+                thread = Thread(target=step_ex.stop)
+                thread.start()
+                threads.append(thread)
+
+            for thread in threads:
+                thread.join()
+
+        self.finish_time = timezone.now()
+        self.state = st.STOPPED
+        logger.logger.info("stage execution stopped", stage_execution_id=self.model.id)
+
     def execute(self) -> None:
         """
         Check if all requirements for execution are met, get init steps and execute them.
         :return: None
         """
         logger.logger.debug("Executing Stage", stage_execution_id=self.model.id)
-        st.StageStateMachine(self.model.id).validate_state(self.state, st.STAGE_EXECUTE_STATES)
+        st.StageStateMachine.validate_state(self.state, st.STAGE_EXECUTE_STATES)
 
         # Stop the execution if dependencies aren't finished.
         if not self.all_dependencies_finished:
@@ -338,7 +291,7 @@ class StageExecution:
         # Get initial Steps in Stage
         step_executions = []
         for step_ex_model in self.model.step_executions.filter(state=st.PENDING, step__is_init=True):
-            step_executions.append(StepExecution(step_execution_id=step_ex_model.id))
+            step_executions.append(StepExecution(step_ex_model.id))
 
         # Pause waiting and awaiting StageExecutions if PlanExecution isn't running.
         if self.state in [st.WAITING, st.AWAITING] and self.model.plan_execution.state != st.RUNNING:
@@ -355,12 +308,42 @@ class StageExecution:
 
         self._run_step_executions(step_executions)
 
-        logger.logger.info(
-            "Stage execution executed",
-            stage_execution_id=self.model.id,
-            stage_name=self.model.stage.name,
-            status="success",
-        )
+        logger.logger.info("stage execution executed", stage_execution_id=self.model.id)
+
+    def pause(self):
+        """
+        Pause stage execution.
+        :return: None
+        """
+        logger.logger.info("stage execution pausing", stage_execution_id=self.model.id)
+        st.StageStateMachine.validate_state(self.state, st.STAGE_PAUSE_STATES)
+        if self.state in [st.AWAITING]:
+            try:
+                self.trigger.stop(trigger_id=self.trigger_id, queue=self.control_queue)
+            except Exception as ex:
+                logger.logger.warning("stage execution cannot be paused", exception=str(ex))
+                return
+            self.pause_time = timezone.now()
+
+        # If stage is RUNNING, set PAUSING state. It will be PAUSED by the currently RUNNING step(s)
+        elif self.state == st.RUNNING:
+            self.state = st.PAUSING
+
+    def resume(self):
+        logger.logger.info("stage execution resuming", stage_execution_id=self.model.id)
+        st.StageStateMachine.validate_state(self.state, st.STAGE_RESUME_STATES)
+
+        self.state = st.RUNNING
+        self.pause_time = None
+
+        if self.all_steps_finished:
+            self.finish()
+            return
+
+        # TODO: pass new_start_time for the trigger
+        #  self.stage_execution.model.plan_execution.start_time - self.stage_execution.pause_time
+        for step_exec in self.model.step_executions.filter(state=st.PAUSED):
+            StepExecution(step_exec.id).start()
 
     def validate_modules(self) -> None:
         """
@@ -368,11 +351,11 @@ class StageExecution:
         """
         logger.logger.debug("Validating stage modules", stage_execution_id=self.model.id)
         for step_ex_id in self.model.step_executions.values_list("id", flat=True):
-            StepExecution(step_execution_id=step_ex_id).validate()
+            StepExecution(step_ex_id).validate()
 
     def report(self) -> dict:
         logger.logger.debug("Generating Stage report", stage_execution_id=self.model.id)
-        report_obj = StageReport(
+        report_obj = dict(
             id=self.model.id,
             name=self.model.stage.name,
             metadata=self.model.stage.metadata,
@@ -384,65 +367,24 @@ class StageExecution:
             step_executions=[],
         )
 
-        for step_execution_obj in StepExecutionModel.objects.filter(stage_execution_id=self.model.id).order_by("id"):
-            step_ex_report = StepExecution(step_execution_id=step_execution_obj.id).report()
-            report_obj.step_executions.append(step_ex_report)
+        for step_execution_model in self.model.step_executions.order_by("id"):
+            step_execution_report = StepExecution(step_execution_model.id).report()
+            report_obj["step_executions"].append(step_execution_report)
 
-        return asdict(report_obj)
+        return report_obj
 
-    def kill(self) -> None:
-        """
-        Kill current StageExecution and its StepExecutions.
-        :return: None
-        """
-        logger.logger.debug("Killing Stage", stage_execution_id=self.model.id)
-        st.StageStateMachine(self.model.id).validate_state(self.state, st.STAGE_KILL_STATES)
-
-        state_before = self.state
-        self.state = st.TERMINATING
-
-        if state_before == st.AWAITING:
-            self.trigger.stop()
-
-        elif state_before == st.WAITING:
-            pass
-
-        else:
-            threads = list()
-
-            # Kill Step executions
-            for step_ex_model in self.model.step_executions.filter(state__in=st.STEP_KILL_STATES):
-                step_ex = StepExecution(step_execution_id=step_ex_model.id)
-                thread = Thread(target=step_ex.kill)
-                threads.append(thread)
-
-            for thread in threads:
-                thread.start()
-
-            for thread in threads:
-                thread.join()
-
-        self.finish_time = timezone.now()
-        self.state = st.TERMINATED
-        logger.logger.info(
-            "Stage execution killed",
-            stage_execution_id=self.model.id,
-            stage_name=self.model.stage.name,
-            status="success",
-        )
-
-    def execute_subjects_to_dependency(self) -> None:
+    def _execute_subjects_to_dependency(self) -> None:
         """
         Execute WAITING StageExecution subjects to specified StageExecution dependency.
         :return: None
         """
         logger.logger.debug("Executing Stage dependency subjects", stage_execution_id=self.model.id)
         subject_to_ids = self.model.stage.subjects_to.all().values_list("stage_id", flat=True)
-        subject_to_exs = self.filter(
-            stage_id__in=subject_to_ids, plan_execution_id=self.model.plan_execution_id, state=st.WAITING
+        subject_to_exs = self.model.plan_execution.stage_executions.filter(
+            stage_id__in=subject_to_ids, state=st.WAITING
         )
         for subject_to_ex in subject_to_exs:
-            subject_to_ex_obj = StageExecution(stage_execution_id=subject_to_ex.id)
+            subject_to_ex_obj = StageExecution(subject_to_ex.id)
             Thread(target=subject_to_ex_obj.execute).run()
 
     def re_execute(self, immediately: bool = False) -> None:
@@ -450,20 +392,20 @@ class StageExecution:
         Reset execution data and re-execute StageExecution.
         :return: None
         """
-        st.StageStateMachine(self.model.id).validate_state(self.state, st.STAGE_FINAL_STATES)
-        self.reset_execution_data()
+        st.StageStateMachine.validate_state(self.state, st.STAGE_FINAL_STATES)
+        self._reset_execution_data()
         if immediately:
             self.execute()
         else:
             self.trigger.start()
 
-    def reset_execution_data(self) -> None:
+    def _reset_execution_data(self) -> None:
         """
         Reset changeable data to defaults and reset StepExecutions.
         :return: None
         """
         logger.logger.debug("Resetting Stage execution data", stage_execution_id=self.model.id)
-        st.StageStateMachine(self.model.id).validate_state(self.state, st.STAGE_FINAL_STATES)
+        st.StageStateMachine.validate_state(self.state, st.STAGE_FINAL_STATES)
 
         with transaction.atomic():
             model = self.model
@@ -479,19 +421,19 @@ class StageExecution:
             model.save()
 
         for step_ex_model in self.model.step_executions.all():
-            StepExecution(step_execution_id=step_ex_model.id).reset_execution_data()
+            StepExecution(step_ex_model.id).reset_execution_data()
 
     def finish(self) -> None:
         """
         Finish execution - update necessary variables, execute dependencies, and stop trigger.
         :return: None
         """
-        st.StageStateMachine(self.model.id).validate_state(self.state, [st.PAUSING, st.RUNNING])
+        st.StageStateMachine.validate_state(self.state, [st.PAUSING, st.RUNNING])
         logger.logger.info("Stage execution finished", stage_execution_id=self.model.id)
 
         self.state = st.FINISHED
         self.finish_time = timezone.now()
-        self.execute_subjects_to_dependency()
+        self._execute_subjects_to_dependency()
 
 
 def execution(execution_id: int) -> None:
@@ -501,5 +443,4 @@ def execution(execution_id: int) -> None:
     :return: None
     """
     logger.logger.debug("Starting Stage execution", stage_execution_id=execution_id)
-    StageExecution(stage_execution_id=execution_id).execute()
-    return None
+    StageExecution(execution_id).execute()
