@@ -1,52 +1,33 @@
 import json
 from datetime import datetime
-from typing import Union, Type, Optional
+from typing import Type
 import re
 import copy
 import yaml
 import amqpstorm
 from jinja2 import nativetypes, StrictUndefined, UndefinedError
 
-from django.core.exceptions import ObjectDoesNotExist
 from django.db import transaction
-from django.db.models.query import QuerySet
-from django.core import exceptions as django_exc
 from django.utils import timezone
 
 from cryton.hive.cryton_app.models import (
     StepModel,
     StepExecutionModel,
-    SuccessorModel,
-    ExecutionVariableModel,
     CorrelationEventModel,
     StepOutputSettingsModel,
     OutputMappingModel,
     StageExecutionModel,
+    SuccessorModel,
 )
 
+from cryton.hive.models.abstract import Instance, Execution
 from cryton.hive.config.settings import SETTINGS
 from cryton.hive.utility import constants, exceptions, logger, states, util, rabbit_client
 from cryton.hive.models import worker
 from cryton.lib.utility.enums import Result
 
-from dataclasses import dataclass, asdict
 
-
-@dataclass
-class StepReport:
-    id: int
-    name: str
-    metadata: dict
-    state: str
-    start_time: datetime
-    finish_time: datetime
-    output: str
-    serialized_output: dict
-    result: str
-    valid: bool
-
-
-class Step:
+class Step(Instance):
     def __init__(self, model_id: int):
         """
         :param model_id: Model ID
@@ -88,7 +69,7 @@ class Step:
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[StepModel], StepModel]:
+    def model(self) -> StepModel:
         self.__model.refresh_from_db()
         return self.__model
 
@@ -96,132 +77,43 @@ class Step:
     def name(self) -> str:
         return self.model.name
 
-    @name.setter
-    def name(self, value: str):
-        model = self.model
-        model.name = value
-        model.save()
-
     @property
     def is_init(self) -> bool:
         return self.model.is_init
-
-    @is_init.setter
-    def is_init(self, value: bool):
-        model = self.model
-        model.is_init = value
-        model.save()
 
     @property
     def arguments(self) -> dict:
         return self.model.arguments
 
-    @arguments.setter
-    def arguments(self, value: dict):
-        model = self.model
-        model.arguments = value
-        model.save()
-
     @property
     def meta(self) -> dict:
         return self.model.metadata
 
-    @meta.setter
-    def meta(self, value: dict):
-        model = self.model
-        model.metadata = value
-        model.save()
 
-    @property
-    def execution_stats_list(self) -> QuerySet:
+class StepExecution(Execution):
+    def __init__(self, model_id: int):
         """
-        Returns StepExecutionStatsModel QuerySet. If the latest is needed, use '.latest()' on result.
-        :return: QuerySet of StepExecutionStatsModel
+        :param model_id: Model ID
         """
-        return StepExecutionModel.objects.filter(step_id=self.model.id)
-
-    @property
-    def parents(self) -> QuerySet:
-        return StepModel.objects.filter(
-            id__in=SuccessorModel.objects.filter(successor_id=self.model.id).values_list("parent_id")
-        )
-
-    @property
-    def successors(self) -> QuerySet:
-        return StepModel.objects.filter(
-            id__in=SuccessorModel.objects.filter(parent_id=self.model.id).values_list("successor_id")
-        )
+        self.__model = StepExecutionModel.objects.get(id=model_id)
 
     @staticmethod
-    def filter(**kwargs) -> QuerySet:
-        """
-        Get list of StepInstances according to no or specified conditions
-        :param kwargs: dict of parameters to filter by
-        :return:
-        """
-        if kwargs:
-            try:
-                return StepModel.objects.filter(**kwargs)
-            except django_exc.FieldError as ex:
-                raise exceptions.WrongParameterError(message=ex)
-        else:
-            return StepModel.objects.all()
+    def create_model(step_id: int, stage_execution_id: int):
+        return StepExecutionModel.objects.create(step_id=step_id, stage_execution_id=stage_execution_id)
 
-    def add_successor(self, successor_id: int, successor_type: str, successor_value: Optional[str]) -> int:
-        """
-        Check if successor's parameters are correct and save it.
-        :param successor_id:
-        :param successor_type: One of valid types
-        :param successor_value: One of valid values for specified type
-        :raises:
-            InvalidSuccessorType
-            InvalidSuccessorValue
-        :return: SuccessorModel id
-        """
-        if successor_type == constants.ANY and successor_value is None:
-            successor_value = ""
+    @classmethod
+    def prepare(cls, step_id: int, stage_execution_id: int) -> "StepExecution":
+        model = cls.create_model(step_id, stage_execution_id)
 
-        successor = SuccessorModel(
-            parent_id=self.model.id, successor_id=successor_id, type=successor_type, value=successor_value
-        )
-        successor.save()
-
-        logger.logger.debug("Step successor created", step_id=self.model.id, step_successor_id=successor_id)
-
-        return successor.id
-
-
-class StepExecution:
-
-    def __init__(self, **kwargs):
-        """
-        :param kwargs:
-        (optional) step_execution_id: int - for retrieving existing execution
-        step_id: int - for creating new execution
-        """
-        step_execution_id = kwargs.get("step_execution_id")
-        if step_execution_id is not None:
-            try:
-                self.model = StepExecutionModel.objects.get(id=step_execution_id)
-            except django_exc.ObjectDoesNotExist:
-                raise exceptions.StepExecutionObjectDoesNotExist(
-                    "StepExecutionStatsModel with id {} does not exist.".format(step_execution_id)
-                )
-
-        else:
-            self.model = StepExecutionModel.objects.create(**kwargs)
+        return StepExecution(model.id)
 
     def delete(self):
         self.model.delete()
 
     @property
-    def model(self) -> Union[Type[StepExecutionModel], StepExecutionModel]:
+    def model(self) -> Type[StepExecutionModel] | StepExecutionModel:
         self.__model.refresh_from_db()
         return self.__model
-
-    @model.setter
-    def model(self, value: StepExecutionModel):
-        self.__model = value
 
     @property
     def state(self) -> str:
@@ -231,26 +123,13 @@ class StepExecution:
     def state(self, value: str):
         with transaction.atomic():
             StepExecutionModel.objects.select_for_update().get(id=self.model.id)
-            if states.StepStateMachine(self.model.id).validate_transition(self.state, value):
+            if states.StepStateMachine.validate_transition(self.state, value):
                 logger.logger.debug(
-                    "Step execution changed state",
-                    step_execution_id=self.model.id,
-                    state_from=self.state,
-                    state_to=value,
+                    "state updated", step_execution_id=self.model.id, state_from=self.state, state_to=value
                 )
                 model = self.model
                 model.state = value
                 model.save()
-
-    @property
-    def result(self) -> str:
-        return self.model.result
-
-    @result.setter
-    def result(self, value: str):
-        model = self.model
-        model.result = value
-        model.save()
 
     @property
     def output(self) -> str:
@@ -263,31 +142,31 @@ class StepExecution:
         model.save()
 
     @property
-    def serialized_output(self) -> Union[list, dict]:
+    def serialized_output(self) -> list | dict:
         return self.model.serialized_output
 
     @serialized_output.setter
-    def serialized_output(self, value: Union[list, dict]):
+    def serialized_output(self, value: list | dict):
         model = self.model
         model.serialized_output = value
         model.save()
 
     @property
-    def start_time(self) -> Optional[datetime]:
+    def start_time(self) -> datetime | None:
         return self.model.start_time
 
     @start_time.setter
-    def start_time(self, value: Optional[datetime]):
+    def start_time(self, value: datetime | None):
         model = self.model
         model.start_time = value
         model.save()
 
     @property
-    def finish_time(self) -> Optional[datetime]:
+    def finish_time(self) -> datetime | None:
         return self.model.finish_time
 
     @finish_time.setter
-    def finish_time(self, value: Optional[datetime]):
+    def finish_time(self, value: datetime | None):
         model = self.model
         model.finish_time = value
         model.save()
@@ -312,28 +191,108 @@ class StepExecution:
         model.parent = value
         model.save()
 
-    @staticmethod
-    def filter(**kwargs) -> QuerySet:
+    def start(self, rabbit_channel: amqpstorm.Channel = None) -> None:
         """
-        Get list of StepExecutionStatsModel according to specified conditions.
-        :param kwargs: dict of parameters to filter by
-        :return: Desired QuerySet
+        Start step execution.
+        :param rabbit_channel: Rabbit channel to use
+        :return: None
         """
-        if kwargs:
-            try:
-                return StepExecutionModel.objects.filter(**kwargs)
-            except django_exc.FieldError as ex:
-                raise exceptions.WrongParameterError(message=ex)
+        logger.logger.debug("starting step execution", step_execution_id=self.model.id)
+        states.StepStateMachine.validate_state(self.state, states.STEP_EXECUTE_STATES)
+
+        self.start_time = timezone.now()
+        self.state = states.STARTING
+
+        try:
+            module_arguments = self.update_step_arguments()
+        except exceptions.MissingValueError as ex:
+            self.state = states.ERROR
+            self.output = str(ex)
+            self.process_error_state()
+            return
+
+        message_body = {constants.MODULE: self.model.step.module, constants.ARGUMENTS: module_arguments}
+        target_queue = worker.Worker(self.model.stage_execution.plan_execution.worker.id).attack_queue
+        reply_queue = SETTINGS.rabbit.queues.attack_response
+        self._execute_on_worker(rabbit_channel, message_body, reply_queue, target_queue)
+        logger.logger.info("step execution started", step_execution_id=self.model.id)
+
+    def stop(self) -> None:
+        """
+        Stop step execution.
+        :return: None
+        """
+        logger.logger.debug("stopping step", step_execution_id=self.model.id)
+        states.StepStateMachine.validate_state(self.state, states.STEP_STOP_STATES)
+
+        state_before = self.state
+        self.state = states.STOPPING
+
+        if state_before == states.PAUSED:
+            self.finish_time = timezone.now()
+            self.state = states.STOPPED
         else:
-            return StepExecutionModel.objects.all()
+            worker_obj = worker.Worker(self.model.stage_execution.plan_execution.worker.id)
+            correlation_id = self.model.correlation_events.first().correlation_id
+            message = {
+                constants.EVENT_T: constants.EVENT_STOP_STEP_EXECUTION,
+                constants.EVENT_V: {"correlation_id": correlation_id},
+            }
+
+            with rabbit_client.RpcClient() as rpc_client:
+                try:
+                    response = rpc_client.call(worker_obj.control_queue, message)
+                except exceptions.RpcTimeoutError as ex:
+                    logger.logger.error("cannot stop step execution", step_execution_id=self.model.id, error=str(ex))
+                    return
+                else:  # The state is set in the `post_process` method
+                    response_value = response.get(constants.EVENT_V)
+                    if response_value.get(constants.RESULT) != Result.OK:
+                        logger.logger.warning(
+                            "cannot stop step execution",
+                            step_execution_id=self.model.id,
+                            error=response_value.get("output"),
+                        )
+                        return
+
+        logger.logger.info("step execution stopped", step_execution_id=self.model.id)
+
+    def finish(self):
+        self._process_results()
+        self._process_successors()
+
+    def _process_results(self):
+        pass
+
+    def _process_successors(self):
+        pass
+
+    def report(self) -> dict:
+        """
+        Generate report containing output from Step Execution.
+        :return: Step Execution report
+        """
+        report_obj = dict(
+            id=self.model.id,
+            name=self.model.step.name,
+            metadata=self.model.step.metadata,
+            state=self.state,
+            start_time=self.start_time,
+            finish_time=self.finish_time,
+            serialized_output=self.serialized_output,
+            output=self.output,
+            valid=self.valid,
+        )
+
+        return report_obj
 
     def validate(self) -> bool:
         """
-        Validate cryton attack module arguments.
+        Validate module arguments.
         :return:
         """
-        logger.logger.debug("Validating Cryton module", step_id=self.model.step.id)
-        target_queue = worker.Worker(self.model.stage_execution.plan_execution.worker.id).control_q_name
+        logger.logger.debug("validate module", step_execution_id=self.model.id)
+        target_queue = worker.Worker(self.model.stage_execution.plan_execution.worker.id).control_queue
         message = {
             constants.EVENT_T: constants.EVENT_VALIDATE_MODULE,
             constants.EVENT_V: {
@@ -350,25 +309,6 @@ class StepExecution:
 
         return self.valid
 
-    def save_output(self, step_output: dict) -> None:
-        """
-        Save Step execution output to StepExecutionModel.
-        :param step_output: dictionary with keys: output, serialized_output
-        :return: None
-        """
-        if (serialized_output := step_output.get(constants.SERIALIZED_OUTPUT)) is not None:
-            for mapping in self.model.step.output_settings.mappings.all():
-                util.rename_key(serialized_output, mapping.name_from, mapping.name_to)
-
-        model = self.model
-
-        if serialized_output is not None:
-            model.serialized_output = serialized_output
-        if (output := step_output.get(constants.OUTPUT)) is not None:
-            model.output = output
-
-        model.save()
-
     def _update_dynamic_variables(self, arguments: dict) -> dict:
         """
         Update dynamic variables in mod_args (even with special $parent prefix)
@@ -382,7 +322,8 @@ class StepExecution:
         prefixes = util.get_prefixes(variable_definitions, dynamic_variable_separator)
         variables: dict[str, dict] = dict()
         executions = StepExecutionModel.objects.filter(
-            stage_execution__plan_execution=self.model.stage_execution.plan_execution
+            stage_execution__plan_execution=self.model.stage_execution.plan_execution,
+            state__in=states.STEP_FINAL_STATES,
         ).order_by("finish_time")
         for prefix in prefixes:
             if prefix == "parent":
@@ -403,7 +344,7 @@ class StepExecution:
         return util.fill_dynamic_variables(copy.deepcopy(arguments), variables, dynamic_variable_separator)
 
     @staticmethod
-    def _update_arguments_with_execution_variables(arguments: dict, execution_variables: list) -> dict:
+    def _update_execution_variables(arguments: dict, execution_variables: list) -> dict:
         """
         Fill Jinja variables in the arguments with execution variables.
         :param arguments: Arguments to fill
@@ -428,24 +369,18 @@ class StepExecution:
         except (yaml.YAMLError, UndefinedError, TypeError) as ex:
             raise exceptions.StepValidationError(f"An error occurred while updating execution variables. {ex}.")
 
-    def update_step_arguments(self, arguments: dict, plan_execution_id: Type[int]) -> dict:
+    def update_step_arguments(self) -> dict:
         """
         Update Step arguments with execution and dynamic variables.
-        :param arguments: Arguments to be updated
-        :param plan_execution_id: ID of the parent step of the current step execution
-        :return: Updated arguments with execution and dynamic variables
+        :return: Arguments updated with execution and dynamic variables
         """
-        execution_vars = list(ExecutionVariableModel.objects.filter(plan_execution_id=plan_execution_id).values())
+        arguments = self.model.step.arguments
+        execution_vars = list(self.model.stage_execution.plan_execution.execution_variables.all().values())
 
         if execution_vars:
-            logger.logger.debug(
-                "Updating step arguments with execution variables",
-                step_execution_id=self.model.id,
-                step_arguments=arguments,
-                execution_vars=execution_vars,
-            )
+            logger.logger.debug("Updating step arguments with execution variables", step_execution_id=self.model.id)
             try:
-                arguments.update(self._update_arguments_with_execution_variables(arguments, execution_vars))
+                arguments.update(self._update_execution_variables(arguments, execution_vars))
             except exceptions.StepValidationError as ex:
                 logger.logger.error(str(ex))
                 raise exceptions.MissingValueError(ex)
@@ -457,10 +392,10 @@ class StepExecution:
             logger.logger.error(str(ex))
             raise exceptions.MissingValueError(f"Failed to update the dynamic variables. Original error: {ex}")
 
-        logger.logger.debug("Step arguments updated", step_execution_id=self.model.id, step_arguments=arguments)
+        logger.logger.debug("Step arguments updated", step_execution_id=self.model.id, arguments=arguments)
         return arguments
 
-    def send_step_execution_request(
+    def _execute_on_worker(
         self, rabbit_channel: amqpstorm.Channel, message_body: dict, reply_queue: str, target_queue: str
     ) -> None:
         """
@@ -471,119 +406,58 @@ class StepExecution:
         :param target_queue: Queue on which should data be sent to worker(for example)
         :return: None
         """
-        logger.logger.debug(
-            "Sending Step execution request", step_execution_id=self.model.id, message_body=message_body
-        )
+        logger.logger.debug("execute step on worker", step_execution_id=self.model.id, message_body=message_body)
         with rabbit_client.RpcClient(rabbit_channel) as rpc_client:
             try:
                 response = rpc_client.call(target_queue, message_body, custom_reply_queue=reply_queue)
             except exceptions.RpcTimeoutError as ex:
                 self.state = states.ERROR
                 self.process_error_state()
-                logger.logger.error("Step execution failed", step_execution_id=self.model.id, error=str(ex))
+                logger.logger.error("step execution failed", step_execution_id=self.model.id, error=str(ex))
                 return
 
             self.state = states.RUNNING
-            CorrelationEventModel.objects.create(
-                correlation_id=rpc_client.correlation_id, step_execution_id=self.model.id
-            )
-            logger.logger.debug(
-                "Received response from Step execution request ", step_execution_id=self.model.id, response=response
-            )
+            CorrelationEventModel.objects.create(correlation_id=rpc_client.correlation_id, step_execution=self.model)
+            logger.logger.debug("step executed on worker", step_execution_id=self.model.id, response=response)
 
-    def _prepare_execution(self) -> [Type[int], worker.Worker]:
+    def _is_successor_executable(self, successor: SuccessorModel) -> bool:
         """
-        Execute necessary actions and return variables needed for individual execution of each type of StepExecution.
-        :return: Plan execution id and worker instance for current Plan execution
+        Check if a successor is runnable by the current execution.
+        :param successor: Possible successor
+        :return: True in case the successor is eligible for execution
         """
-        states.StepStateMachine(self.model.id).validate_state(self.state, states.STEP_EXECUTE_STATES)
-        logger.logger.debug("Executing Step", step_execution_id=self.model.id)
+        match successor.type:
+            case constants.ANY:
+                return True
+            case constants.STATE:
+                if successor.value.lower() == self.state.lower():
+                    return True
+            case constants.OUTPUT:
+                if re.search(successor.value, self.output):
+                    return True
+            case constants.SERIALIZED_OUTPUT:
+                if re.search(successor.value, str(self.serialized_output)):
+                    return True
 
-        plan_execution_id = self.model.stage_execution.plan_execution_id
-        step_worker_obj = self.model.stage_execution.plan_execution.worker
-        worker_obj = worker.Worker(step_worker_obj.id)
+        return False
 
-        # Set STARTING state
-        self.start_time = timezone.now()
-        self.state = states.STARTING
-
-        return plan_execution_id, worker_obj
-
-    def execute(self, rabbit_channel: amqpstorm.Channel = None) -> None:
+    def _get_executable_successors(self) -> set["StepExecution"]:
         """
-        Execute Step on worker specified in execution stats.
-        :param rabbit_channel: Rabbit channel
-        :return: None
+        Get Successors that will be executed by the current execution.
+        :return: Successors
         """
-        logger.logger.debug("Starting Step execution", step_execution_id=self.model.id, step_name=self.model.step.name)
-        plan_execution_id, worker_obj = self._prepare_execution()
+        logger.logger.debug("get executable successors", step_execution_id=self.model.id)
+        successors: set[StepExecution] = set()
+        for successor_model in self.model.step.successors.all():
+            if not self._is_successor_executable(successor_model):
+                continue
 
-        try:
-            module_arguments = self.update_step_arguments(self.model.step.arguments, plan_execution_id)
-        except exceptions.MissingValueError as ex:
-            self.state = states.ERROR
-            self.output = str(ex)
-            self.process_error_state()
-            return
+            step = successor_model.successor
+            step_execution = self.model.stage_execution.step_executions.filter(step=step, state=states.PENDING).first()
+            if step_execution:
+                successors.add(StepExecution(step_execution.id))
 
-        message_body = {
-            constants.MODULE: self.model.step.module,
-            constants.ARGUMENTS: module_arguments,
-        }
-
-        target_queue = worker_obj.attack_q_name
-        reply_queue = SETTINGS.rabbit.queues.attack_response
-
-        self.send_step_execution_request(rabbit_channel, message_body, reply_queue, target_queue)
-        logger.logger.info(
-            "Step executed", step_execution_id=self.model.id, step_name=self.model.step.name, status="success"
-        )
-
-    def report(self) -> dict:
-        """
-        Generate report containing output from Step Execution.
-        :return: Step Execution report
-        """
-        report_obj = StepReport(
-            id=self.model.id,
-            name=self.model.step.name,
-            metadata=self.model.step.metadata,
-            state=self.state,
-            start_time=self.start_time,
-            finish_time=self.finish_time,
-            result=self.result,
-            serialized_output=self.serialized_output,
-            output=self.output,
-            valid=self.valid,
-        )
-
-        return asdict(report_obj)
-
-    def get_successors_to_execute(self) -> QuerySet:
-        """
-        Get Successors based on evaluated dependency.
-
-        :return: QuerySet of StepModel objects
-        """
-        # Get step successor from DB
-        successor_ids = set()
-
-        logger.logger.debug("Getting Step execution successors", step_execution_id=self.model.id)
-        for successor in self.model.step.successors.all():
-            successor_type = successor.type
-            parent_value = getattr(self, successor_type) if successor_type != constants.ANY else None
-            if successor_type == constants.ANY:
-                successor_ids.add(successor.successor_id)
-
-            elif successor_type in constants.REGEX_TYPES:
-                if re.search(successor.value, str(parent_value)):
-                    successor_ids.add(successor.successor_id)
-
-            else:
-                if successor.value == parent_value:
-                    successor_ids.add(successor.successor_id)
-
-        return StepModel.objects.filter(id__in=successor_ids)
+        return successors
 
     def ignore(self) -> None:
         """
@@ -591,35 +465,28 @@ class StepExecution:
 
         :return: None
         """
-        # Stop recursion
         logger.logger.debug("Ignoring Step execution", step_execution_id=self.model.id)
         if self.state == states.IGNORED:
-            return None
-        # If any non SKIPPED parent exists (ignoring the one that called ignore())
-        for parent_step in Step(self.model.step.id).parents:
-            parent_step_exec_model = StepExecutionModel.objects.get(
-                step=parent_step, stage_execution=self.model.stage_execution
-            )
-            parent_step_exec_obj = StepExecution(step_execution_id=parent_step_exec_model.id)
-            # TODO: condition after 'or' may be adding too much runtime complexity here
-            if (
-                parent_step_exec_model.state not in states.STEP_FINAL_STATES
-                or self.model.step in parent_step_exec_obj.get_successors_to_execute()
-            ):
-                return None
+            return
 
-        # Set ignore state
+        # If any non SKIPPED parent exists (ignoring the one that called ignore())
+        if self.model.parent:
+            return
+
+        if (
+            self.model.stage_execution.step_executions.filter(step__in=self.model.step.parents.values("parent"))
+            .exclude(state__in=states.STEP_FINAL_STATES)
+            .exists()
+        ):
+            return
+
         self.state = states.IGNORED
         logger.logger.debug("Step execution ignored", step_execution_id=self.model.id)
-        # Execute for all successors
-        for successor_step in Step(self.model.step.id).successors:
-            step_ex_id = StepExecutionModel.objects.get(
-                step=successor_step, stage_execution=self.model.stage_execution
-            ).id
-            step_ex_obj = StepExecution(step_execution_id=step_ex_id)
+        # Try to ignore successors
+        for successor_step in self.model.step.successors.all():
+            step_ex_id = self.model.stage_execution.step_executions.get(step=successor_step).id
+            step_ex_obj = StepExecution(step_ex_id)
             step_ex_obj.ignore()
-
-        return None
 
     def postprocess(self, ret_vals: dict) -> None:
         """
@@ -630,40 +497,36 @@ class StepExecution:
         :return: None
         """
         logger.logger.debug("Postprocessing Step execution", step_execution_id=self.model.id)
-        # Set final state, optionally save result
-        result = ret_vals.get(constants.RESULT)
         self.finish_time = timezone.now()
 
-        if result == Result.TERMINATED:
-            self.state = states.TERMINATED
-        elif result == Result.ERROR:
-            self.state = states.ERROR
-        else:
-            self.state = states.FINISHED
-            try:
-                self.result = Result(result)
-            except ValueError:
-                self.result = Result.UNKNOWN
+        serialized_output = ret_vals[constants.SERIALIZED_OUTPUT]
+        output = ret_vals[constants.OUTPUT]
+        result = Result(ret_vals[constants.RESULT])
 
-        if self.state == states.FINISHED:
-            ret_vals[constants.OUTPUT], ret_vals[constants.SERIALIZED_OUTPUT] = self._alter_output(
-                ret_vals.get(constants.OUTPUT), ret_vals.get(constants.SERIALIZED_OUTPUT)
-            )
+        match result:
+            case Result.STOPPED:
+                self.state = states.STOPPED
+            case Result.ERROR:
+                self.state = states.ERROR
+            case Result.FAIL:
+                self.state = states.FAILED
+            case Result.OK:
+                self.state = states.FINISHED
+                output, serialized_output = self._alter_output(output, serialized_output)
+                self._apply_output_mappings(serialized_output)
 
-        # Store job output and error message
-        self.save_output(ret_vals)
+        self.serialized_output = serialized_output
+        self.output = output
 
         # update Successors parents
-        successor_list = self.get_successors_to_execute()
-
-        for successor_step in successor_list:
-            successor_step_execution_model = StepExecutionModel.objects.get(
-                step_id=successor_step.id, stage_execution_id=self.model.stage_execution_id, state=states.PENDING
-            )
-            StepExecution(step_execution_id=successor_step_execution_model.id).parent = self.model
+        for successor in self._get_executable_successors():
+            successor.parent = self.model
 
         logger.logger.debug("Step execution postprocess finished", step_execution_id=self.model.id)
-        return None
+
+    def _apply_output_mappings(self, serialized_output: dict):
+        for mapping in self.model.step.output_settings.mappings.all():
+            util.rename_key(serialized_output, mapping.name_from, mapping.name_to)
 
     def _alter_output(self, output: str, serialized_output: dict) -> tuple[str, dict]:
         converted_serialized_output = json.dumps(serialized_output)
@@ -682,118 +545,46 @@ class StepExecution:
         :return: None
         """
         logger.logger.debug("Ignoring Step successors", step_execution_id=self.model.id)
-        step_obj = Step(self.model.step.id)
+        all_successors = self.model.stage_execution.step_executions.filter(
+            step__in=self.model.step.successors.values("successor"), state=states.PENDING
+        )
 
-        # Get all possible successors
-        all_successor_list = StepModel.objects.filter(id__in=step_obj.model.successors.all().values_list("successor"))
-
-        if self.state == states.FINISHED:
-            # Get correct step successor from DB which are to be executed
-            successor_list = self.get_successors_to_execute()
-            # Set IGNORE steps (all successors which won't be executed and don't have parents
-            successor_to_be_skipped = all_successor_list.difference(successor_list)
+        if self.state in [states.FINISHED, states.FAILED, states.ERROR]:
+            executable_successors = [successor.model.id for successor in self._get_executable_successors()]
+            successor_to_be_skipped = all_successors.exclude(id__in=executable_successors)
         else:
-            successor_to_be_skipped = all_successor_list
-        for successor_step in successor_to_be_skipped:
-            try:
-                successor_step_exec_id = StepExecutionModel.objects.get(
-                    step_id=successor_step.id, stage_execution=self.model.stage_execution_id, state=states.PENDING
-                ).id
-            except ObjectDoesNotExist:
-                # Does not exist or is not PENDING
-                continue
-            StepExecution(step_execution_id=successor_step_exec_id).ignore()
-        return None
+            successor_to_be_skipped = all_successors
 
-    def execute_successors(self) -> None:
+        for successor in successor_to_be_skipped:
+            StepExecution(successor.id).ignore()
+
+    def start_successors(self) -> None:
         """
-        Execute all successors of current Step Execution.
+        Execute eligible successors.
         :return: None
         """
-        logger.logger.debug("Executing Step successors", step_execution_id=self.model.id)
-
-        # Get correct step successor from DB which are to be executed
-        successor_list = self.get_successors_to_execute()
-
-        # Execute all successors
-        for successor_step_model in successor_list:
-            successor_step_execution_model = StepExecutionModel.objects.get(
-                step_id=successor_step_model.id,
-                stage_execution_id=self.model.stage_execution_id,
-                state=states.PENDING,
-            )
-            successor_step_exec = StepExecution(step_execution_id=successor_step_execution_model.id)
-            successor_step_exec.execute()
-        return None
+        logger.logger.debug("starting successors", step_execution_id=self.model.id)
+        for successor in self._get_executable_successors():
+            successor.start()
 
     def pause_successors(self) -> None:
         """
-        Pause successor Steps of current Step Execution.
+        Set all successors' state to PAUSED, so they can be executed once resumed.
         :return: None
         """
-        logger.logger.debug("Pausing Step successors", step_execution_id=self.model.id)
-        # Set all successors to PAUSED, so they can be recognized/executed when unpaused
-        successor_list = self.get_successors_to_execute()
-        for step_obj in successor_list:
-            successor_exec_id = StepExecutionModel.objects.get(
-                stage_execution=self.model.stage_execution, step_id=step_obj.id
-            ).id
-            StepExecution(step_execution_id=successor_exec_id).state = states.PAUSED
-            logger.logger.info(
-                "Step successor paused", step_execution_id=self.model.id, successor_exec_id=successor_exec_id
-            )
-
-        return None
-
-    def kill(self):
-        """
-        Kill current Step Execution on Worker.
-
-        :return: Dictionary containing return_code and output
-        """
-        logger.logger.debug("Killing Step", step_id=self.model.step.id)
-        states.StepStateMachine(self.model.id).validate_state(self.state, states.STEP_KILL_STATES)
-
-        state_before = self.state
-        self.state = states.TERMINATING
-
-        if state_before == states.PAUSED:
-            self.finish_time = timezone.now()
-            self.state = states.TERMINATED
-        else:
-            worker_obj = worker.Worker(self.model.stage_execution.plan_execution.worker.id)
-            correlation_id = self.model.correlation_events.first().correlation_id
-            message = {
-                constants.EVENT_T: constants.EVENT_KILL_STEP_EXECUTION,
-                constants.EVENT_V: {"correlation_id": correlation_id},
-            }
-
-            with rabbit_client.RpcClient() as rpc_client:
-                try:
-                    response = rpc_client.call(worker_obj.control_q_name, message)
-                except exceptions.RpcTimeoutError as ex:
-                    logger.logger.error("Step execution not terminated", step_execution_id=self.model.id, error=str(ex))
-                    return
-                else:  # The state is set in the `post_process` method
-                    response_value = response.get(constants.EVENT_V)
-                    if response_value.get(constants.RESULT) != Result.OK:
-                        logger.logger.warning(
-                            "Step execution not terminated",
-                            step_execution_id=self.model.id,
-                            error=response_value.get("output"),
-                        )
-                        return
-
-        logger.logger.info("Step execution terminated", step_execution_id=self.model.id)
+        logger.logger.debug("pausing successors", step_execution_id=self.model.id)
+        for successor in self._get_executable_successors():
+            successor.state = states.PAUSED
+            logger.logger.info("successor paused", step_execution_id=self.model.id, successor_id=successor.model.id)
 
     def re_execute(self) -> None:
         """
         Reset execution data and re-execute StepExecution.
         :return: None
         """
-        states.StepStateMachine(self.model.id).validate_state(self.state, states.STEP_FINAL_STATES)
+        states.StepStateMachine.validate_state(self.state, states.STEP_FINAL_STATES)
         self.reset_execution_data()
-        self.execute()
+        self.start()
 
     def reset_execution_data(self) -> None:
         """
@@ -801,7 +592,7 @@ class StepExecution:
         :return: None
         """
         logger.logger.debug("Resetting Step execution data", step_execution_id=self.model.id)
-        states.StepStateMachine(self.model.id).validate_state(self.state, states.STEP_FINAL_STATES)
+        states.StepStateMachine.validate_state(self.state, states.STEP_FINAL_STATES)
 
         with transaction.atomic():
             model = self.model
@@ -811,7 +602,6 @@ class StepExecution:
             model.start_time = None
             model.pause_time = None
             model.finish_time = None
-            model.result = ""
             model.serialized_output = dict()
             model.output = ""
             model.valid = False
