@@ -1,4 +1,4 @@
-from typing import Union, Type, List
+from typing import Type
 from cryton.hive.utility import logger
 
 
@@ -9,7 +9,7 @@ class Error(Exception):
 class ValidationError(Error):
     """Exception raised for errors in the validation process."""
 
-    def __init__(self, message: Union[dict, Exception, str]):
+    def __init__(self, message: dict | Exception | str):
         if type(message) != dict:
             self.message = {"message": message}
         else:
@@ -21,7 +21,7 @@ class ValidationError(Error):
 class PlanValidationError(ValidationError):
     """Exception raised for errors in the Plan validation process."""
 
-    def __init__(self, message: Union[Exception, str], plan_name: str = None):
+    def __init__(self, message: Exception | str, plan_name: str = None):
         self.message = {"message": message, "plan_name": plan_name}
         super().__init__(self.message)
 
@@ -37,7 +37,7 @@ class DuplicateNameInPlan(PlanValidationError):
 class StageValidationError(ValidationError):
     """Exception raised for errors in the Stage validation process."""
 
-    def __init__(self, message: Union[Exception, str], stage_name: str = None):
+    def __init__(self, message: Exception | str, stage_name: str = None):
         self.message = {"message": message, "stage_name": stage_name}
         logger.logger.error(message, stage_name=stage_name, state="fail")
         super().__init__(self.message)
@@ -46,7 +46,7 @@ class StageValidationError(ValidationError):
 class StepValidationError(ValidationError):
     """Exception raised for errors in the Step validation process."""
 
-    def __init__(self, message: Union[Exception, str], step_name: str = None):
+    def __init__(self, message: Exception | str, step_name: str = None):
         self.message = {"message": message, "step_name": step_name}
         super().__init__(self.message)
 
@@ -54,7 +54,7 @@ class StepValidationError(ValidationError):
 class UserInputError(Error):
     """Exception raised for errors when user inputs an invalid input."""
 
-    def __init__(self, message: Union[Exception, str], user_input: Union[str, int] = None):
+    def __init__(self, message: Exception | str, user_input: str | int = None):
         self.message = {"message": message, "user_input": user_input}
         super().__init__(self.message)
 
@@ -91,7 +91,7 @@ class PlanObjectDoesNotExist(ObjectDoesNotExist):
 class PlanExecutionDoesNotExist(ObjectDoesNotExist):
     """Exception raised if Execution does not exist."""
 
-    def __init__(self, message: Union[Exception, str] = "PlanExecution not found", plan_execution_id: str = None):
+    def __init__(self, message: Exception | str = "PlanExecution not found", plan_execution_id: str = None):
         self.message = {"message": message, "plan_execution_id": plan_execution_id}
         super().__init__(self.message)
 
@@ -99,7 +99,7 @@ class PlanExecutionDoesNotExist(ObjectDoesNotExist):
 class StageObjectDoesNotExist(ObjectDoesNotExist):
     """Exception raised if trying to use a Stage object that doesn't exist."""
 
-    def __init__(self, message: Union[Exception, str], stage_id: int = None, plan_id: int = None):
+    def __init__(self, message: Exception | str, stage_id: int = None, plan_id: int = None):
         self.message = {"message": message, "stage_id": stage_id, "plan_id": plan_id}
         super().__init__(self.message)
 
@@ -107,7 +107,7 @@ class StageObjectDoesNotExist(ObjectDoesNotExist):
 class StageExecutionObjectDoesNotExist(ObjectDoesNotExist):
     """Exception raised if trying to use a StageExecution object that doesn't exist."""
 
-    def __init__(self, message: Union[Exception, str], stage_id: int = None, plan_id: int = None):
+    def __init__(self, message: Exception | str, stage_id: int = None, plan_id: int = None):
         self.message = {"message": message, "stage_id": stage_id, "plan_id": plan_id}
         super().__init__(self.message)
 
@@ -117,10 +117,10 @@ class StepObjectDoesNotExist(ObjectDoesNotExist):
 
     def __init__(
         self,
-        message: Union[Exception, str],
+        message: Exception | str,
         step_id: int = None,
         step_name: str = None,
-        stage_id: Union[int, Type[int]] = None,
+        stage_id: int | Type[int] = None,
         plan_id: int = None,
     ):
         self.message = {
@@ -136,7 +136,7 @@ class StepObjectDoesNotExist(ObjectDoesNotExist):
 class StepExecutionObjectDoesNotExist(ObjectDoesNotExist):
     """Exception raised if trying to set invalid type of successor."""
 
-    def __init__(self, message: Union[Exception, str] = "StepExecution not found", step_execution_id: str = None):
+    def __init__(self, message: Exception | str = "StepExecution not found", step_execution_id: str = None):
         self.message = {"message": message, "step_execution_id": step_execution_id}
         super().__init__(self.message)
 
@@ -144,7 +144,7 @@ class StepExecutionObjectDoesNotExist(ObjectDoesNotExist):
 class SuccessorObjectDoesNotExist(ObjectDoesNotExist):
     """Exception raised if trying to use a Successor object that doesn't exist."""
 
-    def __init__(self, message: Union[Exception, str], step_id: int = None):
+    def __init__(self, message: Exception | str, step_id: int = None):
         self.message = {"message": message, "step_id": step_id}
         super().__init__(self.message)
 
@@ -162,10 +162,10 @@ class SessionObjectDoesNotExist(ObjectDoesNotExist):
 
     def __init__(
         self,
-        message: Union[Exception, str],
+        message: Exception | str,
         session_name: str = None,
         session_id: int = None,
-        plan_execution_id: Union[Type[int], int] = None,
+        plan_execution_id: Type[int] | int = None,
         step_id: int = None,
         plan_id: int = None,
         target=None,
@@ -187,9 +187,7 @@ class SessionObjectDoesNotExist(ObjectDoesNotExist):
 class SessionIsNotOpen(Error):
     """Exception raised if the desired session isn't open."""
 
-    def __init__(
-        self, message: Union[Exception, str], session_name: str, plan_execution_id: Union[Type[int], int], step_id: int
-    ):
+    def __init__(self, message: Exception | str, session_name: str, plan_execution_id: Type[int] | int, step_id: int):
         self.message = {
             "message": message,
             "session_name": session_name,
@@ -202,7 +200,7 @@ class SessionIsNotOpen(Error):
 class ArgumentsObjectDoesNotExist(ObjectDoesNotExist):
     """Exception raised if trying to use an Arguments object that doesn't exist."""
 
-    def __init__(self, message: Union[Exception, str], step_id: int = None):
+    def __init__(self, message: Exception | str, step_id: int = None):
         self.message = {"message": message, "step_id": step_id}
         super().__init__(self.message)
 
@@ -210,141 +208,21 @@ class ArgumentsObjectDoesNotExist(ObjectDoesNotExist):
 class InvalidStateError(Error):
     """Exception raised if an invalid state is detected."""
 
-    def __init__(self, message: Union[Exception, str], execution_id: int, state: str, allowed_states: list):
-        self.message = {
-            "message": message,
-            "execution_id": execution_id,
-            "state": state,
-            "allowed_states": allowed_states,
-        }
-        super().__init__(self.message)
-
-
-class RunInvalidStateError(InvalidStateError):
-    """Exception raised if Run's invalid state is detected."""
-
-    def __init__(self, message: Union[Exception, str], execution_id: int, state: str, allowed_states: list):
-        super().__init__(message, execution_id, state, allowed_states)
-        logger.logger.error(message, execution_id=execution_id, state=state, allowed_states=allowed_states)
-
-
-class PlanInvalidStateError(InvalidStateError):
-    """Exception raised if PlanExecution's invalid state is detected."""
-
-    def __init__(self, message: Union[Exception, str], execution_id: int, state: str, allowed_states: list):
-        super().__init__(message, execution_id, state, allowed_states)
-        logger.logger.error(message, execution_id=execution_id, state=state, allowed_states=allowed_states)
-
-
-class StageInvalidStateError(InvalidStateError):
-    """Exception raised if StageExecution's invalid state is detected."""
-
-    def __init__(self, message: Union[Exception, str], execution_id: int, state: str, allowed_states: list):
-        super().__init__(message, execution_id, state, allowed_states)
-        logger.logger.error(message, execution_id=execution_id, state=state, allowed_states=allowed_states)
-
-
-class StepInvalidStateError(InvalidStateError):
-    """Exception raised if StepExecution's invalid state is detected."""
-
-    def __init__(self, message: Union[Exception, str], execution_id: int, state: str, allowed_states: list):
-        super().__init__(message, execution_id, state, allowed_states)
-        logger.logger.error(message, execution_id=execution_id, state=state, allowed_states=allowed_states)
+    def __init__(self, message: Exception | str):
+        super().__init__(message)
 
 
 class StateTransitionError(Error):
     """Exception raised if an invalid state transition is made."""
 
-    def __init__(
-        self,
-        message: Union[Exception, str],
-        execution_id: int,
-        state_from: str,
-        state_to: str,
-        allowed_transitions: list,
-    ):
-        self.message = {
-            "message": message,
-            "execution_id": execution_id,
-            "state_from": state_from,
-            "state_to": state_to,
-            "allowed_transitions": allowed_transitions,
-        }
-        super().__init__(self.message)
-
-
-class RunStateTransitionError(StateTransitionError):
-    """Raised if an invalid Run's state transition is made."""
-
-    def __init__(
-        self,
-        message: Union[Exception, str],
-        execution_id: int,
-        state_from: str,
-        state_to: str,
-        allowed_transitions: list,
-    ):
-        super().__init__(message, execution_id, state_from, state_to, allowed_transitions)
-        logger.logger.warning(
-            "Invalid transition detected in Run", execution_id=execution_id, state_from=state_from, state_to=state_to
-        )
-
-
-class PlanStateTransitionError(StateTransitionError):
-    """Raised if an invalid PlanExecution's state transition is made."""
-
-    def __init__(
-        self,
-        message: Union[Exception, str],
-        execution_id: int,
-        state_from: str,
-        state_to: str,
-        allowed_transitions: list,
-    ):
-        super().__init__(message, execution_id, state_from, state_to, allowed_transitions)
-        logger.logger.warning(
-            "Invalid transition detected in Plan", execution_id=execution_id, state_from=state_from, state_to=state_to
-        )
-
-
-class StageStateTransitionError(StateTransitionError):
-    """Raised if an invalid StageExecution's state transition is made."""
-
-    def __init__(
-        self,
-        message: Union[Exception, str],
-        execution_id: int,
-        state_from: str,
-        state_to: str,
-        allowed_transitions: list,
-    ):
-        super().__init__(message, execution_id, state_from, state_to, allowed_transitions)
-        logger.logger.warning(
-            "Invalid transition detected in Stage", execution_id=execution_id, state_from=state_from, state_to=state_to
-        )
-
-
-class StepStateTransitionError(StateTransitionError):
-    """Raised if an invalid StepExecution's state transition is made."""
-
-    def __init__(
-        self,
-        message: Union[Exception, str],
-        execution_id: int,
-        state_from: str,
-        state_to: str,
-        allowed_transitions: list,
-    ):
-        super().__init__(message, execution_id, state_from, state_to, allowed_transitions)
-        logger.logger.warning(
-            "Invalid transition detected in Step", execution_id=execution_id, state_from=state_from, state_to=state_to
-        )
+    def __init__(self, message: Exception | str):
+        super().__init__(message)
 
 
 class UnexpectedValue(Error):
     """Raised if an invalid value is used."""
 
-    def __init__(self, message: Union[Exception, str], wrong_value: Union[str, int] = None):
+    def __init__(self, message: Exception | str, wrong_value: str | int = None):
         self.message = {"message": message, "wrong_value": wrong_value}
         super().__init__(self.message)
 
@@ -356,7 +234,7 @@ class MissingValueError(Error):
 class WrongParameterError(Error):
     """Exception raised if trying to filter in model by wrong parameter."""
 
-    def __init__(self, message: Union[Exception, str] = "Wrong parameter name", param_name: str = None):
+    def __init__(self, message: Exception | str = "Wrong parameter name", param_name: str = None):
         self.message = {"message": message, "param_name": param_name}
         super().__init__(self.message)
 
@@ -364,7 +242,7 @@ class WrongParameterError(Error):
 class ParameterMissingError(Error):
     """Exception raised if compulsory parameter is missing."""
 
-    def __init__(self, message: Union[Exception, str] = "Missing parameter", param_name: str = None):
+    def __init__(self, message: Exception | str = "Missing parameter", param_name: str = None):
         self.message = {"message": message, "param_name": param_name}
         super().__init__(self.message)
 
@@ -372,7 +250,7 @@ class ParameterMissingError(Error):
 class DependencyDoesNotExist(Error):
     """Exception raised if could not create dependency between Stages."""
 
-    def __init__(self, message: Union[Exception, str], stage_name: str = None):
+    def __init__(self, message: Exception | str, stage_name: str = None):
         self.message = {"message": message, "stage_name": stage_name}
         logger.logger.error("stage dependency does not exist", stage_name=stage_name)
         super().__init__(self.message)
@@ -381,7 +259,7 @@ class DependencyDoesNotExist(Error):
 class InvalidSuccessorType(Error):
     """Exception raised if trying to set invalid type of successor."""
 
-    def __init__(self, message: Union[Exception, str] = "Wrong successor type", successor_type: str = None):
+    def __init__(self, message: Exception | str = "Wrong successor type", successor_type: str = None):
         self.message = {"message": message, "successor_type": successor_type}
         super().__init__(self.message)
 
@@ -389,7 +267,7 @@ class InvalidSuccessorType(Error):
 class InvalidSuccessorValue(Error):
     """Exception raised if trying to set invalid value for successor."""
 
-    def __init__(self, message: Union[Exception, str] = "Wrong successor value", successor_value: str = None):
+    def __init__(self, message: Exception | str = "Wrong successor value", successor_value: str = None):
         self.message = {"message": message, "successor_value": successor_value}
         super().__init__(self.message)
 
@@ -405,7 +283,7 @@ class CreationFailedError(Error):
 class PlanCreationFailedError(CreationFailedError):
     """Exception raised if Plan creation failed."""
 
-    def __init__(self, message: Union[Exception, str], plan_name: str = None):
+    def __init__(self, message: Exception | str, plan_name: str = None):
         self.message = {"message": message, "plan_name": plan_name}
         logger.logger.error("plan creation failed", plan_name=plan_name)
         super().__init__(self.message)
@@ -414,7 +292,7 @@ class PlanCreationFailedError(CreationFailedError):
 class StageCreationFailedError(CreationFailedError):
     """Exception raised if Stage creation failed."""
 
-    def __init__(self, message: Union[Exception, str], stage_name: str = None):
+    def __init__(self, message: Exception | str, stage_name: str = None):
         self.message = {"message": message, "stage_name": stage_name}
         logger.logger.error("stage creation failed", stage_name=stage_name)
         super().__init__(self.message)
@@ -423,7 +301,7 @@ class StageCreationFailedError(CreationFailedError):
 class StepCreationFailedError(CreationFailedError):
     """Exception raised if Step creation failed."""
 
-    def __init__(self, message: Union[Exception, str], step_name: str = None):
+    def __init__(self, message: Exception | str, step_name: str = None):
         self.message = {"message": message, "step_name": step_name}
         logger.logger.error("step creation failed", step_name=step_name)
         super().__init__(self.message)
@@ -432,7 +310,7 @@ class StepCreationFailedError(CreationFailedError):
 class SuccessorCreationFailedError(CreationFailedError):
     """Exception raised if Successor creation failed."""
 
-    def __init__(self, message: Union[Exception, str], successor_name: str = None):
+    def __init__(self, message: Exception | str, successor_name: str = None):
         self.message = {"message": message, "successor_name": successor_name}
         logger.logger.error("Successor creation failed", successor_name=successor_name)
         super().__init__(self.message)
@@ -441,7 +319,7 @@ class SuccessorCreationFailedError(CreationFailedError):
 class PlanExecutionCreationFailedError(CreationFailedError):
     """Exception raised if PlanExecution creation failed."""
 
-    def __init__(self, message: Union[Exception, str] = "Bad argument", param_name: str = None, param_type: str = None):
+    def __init__(self, message: Exception | str = "Bad argument", param_name: str = None, param_type: str = None):
         self.message = {"message": message, "param_type": param_type, "param_name": param_name}
         logger.logger.error("PlanExecution creation failed", param_name=param_name, param_type=param_type)
         super().__init__(self.message)
@@ -450,7 +328,7 @@ class PlanExecutionCreationFailedError(CreationFailedError):
 class RunCreationFailedError(CreationFailedError):
     """Exception raised if Run creation failed."""
 
-    def __init__(self, message: Union[Exception, str]):
+    def __init__(self, message: Exception | str):
         self.message = {"message": message}
         logger.logger.error("run creation failed")
         super().__init__(self.message)
@@ -459,7 +337,7 @@ class RunCreationFailedError(CreationFailedError):
 class RabbitError(Error):
     """Exception raised when there is some problem with RabbitMQ"""
 
-    def __init__(self, message: Union[Exception, str]):
+    def __init__(self, message: Exception | str):
         self.message = {"message": message}
         super().__init__(self.message)
 
@@ -467,14 +345,14 @@ class RabbitError(Error):
 class RpcTimeoutError(RabbitError):
     """Exception raised if the RPC request timeouts"""
 
-    def __init__(self, message: Union[Exception, str]):
+    def __init__(self, message: Exception | str):
         super().__init__(message)
 
 
 class StageCycleDetected(Error):
     """Exception raised if there was Stage cycle detected"""
 
-    def __init__(self, message: Union[Exception, str]):
+    def __init__(self, message: Exception | str):
         self.message = {"message": message}
         super().__init__(self.message)
 
@@ -482,7 +360,7 @@ class StageCycleDetected(Error):
 class TriggerTypeDoesNotExist(Error):
     """Exception raised if the trigger type doesn't exist"""
 
-    def __init__(self, trigger_type: str, supported_triggers: List):
+    def __init__(self, trigger_type: str, supported_triggers: list):
         self.message = {
             "message": f"Nonexistent trigger type '{trigger_type}'. Supported trigger types are: "
             f"{', '.join(supported_triggers)}."
@@ -493,7 +371,7 @@ class TriggerTypeDoesNotExist(Error):
 class StepTypeDoesNotExist(Error):
     """Exception raised if the step type doesn't exist"""
 
-    def __init__(self, step_type: str, supported_step_types: List):
+    def __init__(self, step_type: str, supported_step_types: list):
         self.message = {
             "message": f"Nonexistent step type '{step_type}'. Supported step types are: "
             f"{', '.join(supported_step_types)}."
