@@ -1,4 +1,4 @@
-# Generated by Django 4.2.16 on 2024-09-20 06:43
+# Generated by Django 4.2.16 on 2024-12-02 15:01
 
 from django.db import migrations, models
 import django.db.models.deletion
@@ -193,7 +193,6 @@ class Migration(migrations.Migration):
                 ("finish_time", models.DateTimeField(null=True)),
                 ("serialized_output", models.JSONField(default=dict)),
                 ("output", models.TextField(default="")),
-                ("result", models.TextField(default="")),
                 ("valid", models.BooleanField(default=False)),
                 (
                     "parent",
