@@ -1,6 +1,7 @@
 # types
 RETURN_VALUE = "return_value"
 RESULT = "result"
+STATE = "state"
 OUTPUT = "output"
 SERIALIZED_OUTPUT = "serialized_output"
 ANY = "any"
@@ -22,10 +23,6 @@ TIME = "time"
 HTTP = "http"
 METASPLOIT = "metasploit"
 
-# Successor related
-VALID_SUCCESSOR_TYPES = [RESULT, ANY, SERIALIZED_OUTPUT, OUTPUT]
-SUCCESSOR_TYPES_WITHOUT_ANY = [RESULT, SERIALIZED_OUTPUT, OUTPUT]
-
 # RabbitMQ message keywords
 EVENT_T = "event_t"
 EVENT_V = "event_v"
@@ -43,9 +40,7 @@ SEPARATOR_DEFAULT_VALUE = "."
 
 # Event types
 EVENT_VALIDATE_MODULE = "VALIDATE_MODULE"
-EVENT_LIST_MODULES = "LIST_MODULES"
-EVENT_LIST_SESSIONS = "LIST_SESSIONS"
-EVENT_KILL_STEP_EXECUTION = "KILL_STEP_EXECUTION"
+EVENT_STOP_STEP_EXECUTION = "STOP_STEP_EXECUTION"
 EVENT_HEALTH_CHECK = "HEALTH_CHECK"
 EVENT_ADD_TRIGGER = "ADD_TRIGGER"
 EVENT_REMOVE_TRIGGER = "REMOVE_TRIGGER"
